### 2025-10-13 730547e
Fix doi and orcid id
