+++ paper/paper.md
@@ -10,47 +10,47 @@ tags:
 authors:
   - given-names: Rubel
     surname: Mozumder
-    orcid: https://orcid.org/0009-0007-5926-6646
+    orcid: 0009-0007-5926-6646
     affiliation: 1
   - given-names: Lukas
     surname: Pielsticker
-    orcid: https://orcid.org/0000-0001-9361-8333
+    orcid: 0000-0001-9361-8333
     affiliation: "1,2"
   - given-names: Markus
     surname: Kühbach
-    orcid: https://orcid.org/0000-0002-7117-5196
+    orcid: 0000-0002-7117-5196
     affiliation: 1
   - given-names: Andrea
     surname: Albino
-    orcid: https://orcid.org/0000-0001-9280-7431
+    orcid: 0000-0001-9280-7431
     affiliation: 1
   - given-names: Florian
     surname: Dobener
-    orcid: https://orcid.org/0000-0003-1987-6224
+    orcid: 0000-0003-1987-6224
     affiliation: 1
   - given-names: Sherjeel
     surname: Shabih
-    orcid: https://orcid.org/0009-0008-6635-4465
+    orcid: 0009-0008-6635-4465
     affiliation: 1
   - given-names: José A.
     surname: Márquez Prieto
-    orcid: https://orcid.org/0000-0002-8173-2566
+    orcid: 0000-0002-8173-2566
     affiliation: 1
   - given-names: Sandor
     surname: Brockhauser
-    orcid: https://orcid.org/0000-0002-9700-4803
+    orcid: 0000-0002-9700-4803
     affiliation: 1
   - given-names: Claudia
     surname: Draxl
-    orcid: https://orcid.org/0000-0003-3523-6657
+    orcid: 0000-0003-3523-6657
     affiliation: 1
   - given-names: Christoph
     surname: Koch
-    orcid: https://orcid.org/0000-0002-3984-1523
+    orcid: 0000-0002-3984-1523
     affiliation: 1
   - given-names: Heiko B.
     surname: Weber
-    orcid: https://orcid.org/0000-0002-6403-9022
+    orcid: 0000-0002-6403-9022
     affiliation: 3
 
 affiliations:
