+++ paper/paper.bib
@@ -72,7 +72,7 @@ pages = {151-153},
 year = {1997},
 note = {Proceedings of the International Conference on Neutron Scattering},
 issn = {0921-4526},
-doi = {https://doi.org/10.1016/S0921-4526(97)00865-X},
+doi = {10.1016/S0921-4526(97)00865-X},
 url = {https://www.sciencedirect.com/science/article/pii/S092145269700865X},
 author = {Klosowski, Przemek and K{\"{o}}nnecke, Mark and Tischler, Jonathan Z. and Osborn, Ray},
 }
@@ -83,7 +83,7 @@ volume = {385-386},
 pages = {1343-1345},
 year = {2006},
 issn = {0921-4526},
-doi = {https://doi.org/10.1016/j.physb.2006.06.106},
+doi = {10.1016/j.physb.2006.06.106},
 url = {https://www.sciencedirect.com/science/article/pii/S0921452606013779},
 author = {K{\"{o}}nnecke, Mark},
 }
