### 2026-02-03 1f01c40
Research Impact Statement, Software design, and AI usage disclosure.
