+++ paper/paper.md
@@ -65,17 +65,17 @@ date: 01 July 2025
 bibliography: paper.bib
 ---
 
-# Summary
+# Research Impact Statement
 
 The NeXus scientific data format standard [@Könnecke:2015; @Könnecke2006; @Klosowski1997], which was originally introduced for neutron, X-ray, and muon science, has in recent years seen a significant enhancement across diverse scientific domains such as materials science. NeXus definitions—comprising application definitions and base classes—describe the hierarchical structure and semantics of valid NeXus files. They are written in XML [@xml10] using the NeXus Definition Language (NXDL), which is itself defined in XSD (XML Schema Definition) [@w3c:xmlschema1].
 
-`nyaml` is a Python-based tool with both a command-line interface and an application programming interface (API) that facilitates the conversion between NXDL XML and a simplified YAML [@YAML11] representation. YAML's indentation-based syntax enhances human readability and simplifies manual editing. By providing a reliable, lossless round-trip conversion between XML and YAML, `nyaml` enables developers to edit NeXus definitions efficiently without sacrificing structural or semantic fidelity.
+`nyaml` is a Python-based tool with both a command-line interface and an application programming interface (API) that facilitates the conversion between NXDL XML and a simplified YAML [@YAML11] representation. YAML's indentation-based syntax enhances human readability and simplifies manual editing. By providing a reliable, lossless round-trip conversion between XML and YAML, `nyaml` enables developers to edit NeXus definitions efficiently without sacrificing structural or semantic fidelity. With `nyaml`, scientists who develop and maintain scientific data models in the NeXus framework can achieve a 2-3-fold improvement in efficiency compared to direct XML editing by leverageing the tool's warnings and error messages to avoid common mistakes.
 
 # Statement of need
 
 As the NeXus standard and its community of definition developers continue to grow, it is increasingly important to ensure that the development process remains both user-friendly and resilient to errors. The existing representation of NeXus definitions in XML offers structural rigor through the NeXus Definition Language (NXDL) and a well-defined hierarchy for metadata and data types. However, it is verbose and can be difficult to edit by hand. Writing and maintaining NXDL files often involves dealing with deeply nested elements and strict syntax, which can be error-prone and time-consuming, especially for users who are not familiar with XML development. `nyaml` addresses these challenges by enabling the development of NeXus definitions in a simpler, YAML-based format while preserving the full structure, semantics, and developer comments of the original XML. As a successful stress test, the tool has been used for developing extensions of the NeXus definitions for different microscopy and spectroscopy methods to integrate the definitions into the research data management system NOMAD [@NOMAD2023].
 
-# Conversion process
+# Software design
 
 `nyaml` is a Python package developed for converting NeXus definitions from YAML format (files with a `.yaml` extension) into the XML format (files with a `.nxdl.xml` extension) and vice versa. The package is published on PyPI and therefore can be installed using Python package managers (e.g., `pip` [@pip2025]). The `nyaml` package introduces a set of keywords and syntactic rules (see [documentation](https://fairmat-nfdi.github.io/nyaml)) that are specific to NXDL (NeXus Definition Language) in YAML. The keywords and syntactic rules imply a relationship between YAML and XML (XML tags and attributes) structures, enabling seamless conversion for new and existing definitions. Existing definitions can be modified using the following workflow: **XML $\rightarrow$ YAML $\rightarrow$ modification of YAML $\rightarrow$ XML**. New definitions can be designed in YAML and converted to XML directly
 
@@ -90,6 +90,9 @@ In the reverse direction — **XML to YAML** conversion (see \autoref{fig:nyaml_
 # Evaluation from NIAC
 
[…]
