+++ paper/paper.bib
@@ -123,3 +123,52 @@ pages = {5388},
 title = {NOMAD: A distributed web-based platform for managing materials science research data},
 journal = {Journal of Open Source Software}
 }
+@misc{nyamlFirstCommit2023,
+  author = {{Dobener, F.}},
+  title = {GitHub Commit: Initial commit of nyaml repository},
+  year = {2023},
+  month = {Nov},
+  url = {https://github.com/FAIRmat-NFDI/nyaml/tree/ebc14182aa02fab6dcb862ae844cbf622730c049}
+  }
+@misc{nyamlContributors,
+  author = {{Mozumder, R. et al.}},
+  title = {nyaml GitHub Repository Contributors},
+  year = {2025},
+  month = {Aug},
+  url = {https://github.com/FAIRmat-NFDI/nyaml/graphs/contributors}
+  }
+@misc{NIACOS2025,
+  author = {{Pielsticker, L. and Hildebrandt, Ron  et al.}},
+  title = {Fairmat 2024: proposal on optical spectroscopy},
+  year = {2025},
+  month = {Jul},
+  url = {https://github.com/nexusformat/definitions/pull/1425}
+  }
+@misc{NIACEM2025,
+  author = {{K{\"{u}}hbach, M and Pielsticker, Lukas et al.}},
+  title = {Fairmat 2024: proposal on electron microscopy (EM)},
+  year = {2025},
+  month = {Aug},
+  url = {https://github.com/nexusformat/definitions/pull/1423},
+  }
+@misc{NIACMPES2025,
+  author ={{Dobener, F. and Pielsticker, Lukas et al.}},
+  title = {Fairmat 2024: proposal on photoemission spectroscopy (MPES)},
+  year = {2025},
+  month = {Jun},
+  url = {https://github.com/nexusformat/definitions/pull/1424}
+  }
+@misc{NIACAPM2025,
+  author = {{K{\"{u}}hbach, M. and Pielsticker, Lukas et al.}},
+  title = {Fairmat 2024: proposal on atom probe microscopy (APM) apm proposal},
+  year = {2025},
+  month = {Jul},
+  url = {https://github.com/nexusformat/definitions/pull/1422}
+  }
+@misc{nyamlGH2023,
+  author = {{Mozumder, R. et al.}},
+  title = {nyaml GitHub Repository},
+  year = {2023},
+  month = {Nov},
+  url = {https://github.com/FAIRmat-NFDI/nyaml}
+  }
