### 2026-02-03 67c798f
Includes: Research Impact Statement, Software design, and AI usage disclosure.
