### 2026-02-26 273a530
Fix typos and add missing word in cspell dictionary.
