+++ paper/paper.bib
@@ -99,7 +99,7 @@ author = {K{\"{o}}nnecke, Mark},
   title = {NeXus Definition GitHub Repository},
   year = {2024},
   month = {Feb},
-  url = {https://github.com/nexusformat/definition},
+  url = {https://github.com/nexusformat/definitions},
   release = {2024.02}
 }
 @misc{w3c:xmlschema1,
@@ -149,7 +149,7 @@ journal = {Journal of Open Source Software}
   title = {Fairmat 2024: proposal on electron microscopy (EM)},
   year = {2025},
   month = {Aug},
-  url = {https://github.com/nexusformat/definitions/pull/1423},
+  url = {https://github.com/nexusformat/definitions/pull/1423}
   }
 @misc{NIACMPES2025,
   author ={{Dobener, F. and Pielsticker, Lukas et al.}},
