+++ .cspell/custom-dictionary.txt
@@ -89,7 +89,6 @@ Mozumder
 Mülheim
 NFDI
 NIAC
-NXDL
 Pielsticker
 Physik
 Prieto
@@ -105,3 +104,7 @@ lxml
 nyaml
 orcid
 xmlschema1
+NIACEM
+NIACMPES
+NIACOS
+NIACAPM
