### 2026-03-16 f8b3054
use `yaml.SafeLoader` (#85)
Co-authored-by: lukaspie <lukaspie@github.com>
