### 2026-03-19 #10231 @editorialbot
[REVIEW]: nyaml: Format Converter for the NeXus Data Model
**Submitting author:** @RubelMozumder (<a href="http://orcid.org/0009-0007-5926-6646">Rubel Mozumder</a>)
**Repository:** https://github.com/FAIRmat-NFDI/nyaml
**Branch with paper.md** (empty if default branch): JossPublication
**Version:** v0.2.3
