### 2026-04-13 FAIRmat-NFDI/nyaml#86 @maurov
[JOSS review] Contributing guidelines are missing
Hello,
Related to the [JOSS paper review](https://github.com/openjournals/joss-reviews/issues/10231), this point is missing:

Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Please, could you add a `CONTRIBUTING.md` in the repository or a section in the documentation?
