### 2026-04-13 FAIRmat-NFDI/nyaml#87 @maurov
[JOSS review] Question about testing/validation of YAML files?
Hello,
Within the [JOSS paper review](https://github.com/openjournals/joss-reviews/issues/10231), I was wondering how to improve the following point:

Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I acknowledge the presence of the test suite in your code, based on the XML/YAML files you have provided for your own NeXus definitions, but how `nyaml` could help a user to validate the written YAML in order that the generated XML file would be well suited for submitting to the NIAC?

For a non-expert in NeXus, the documentation on [Writing a NeXus definition in YAML](https://fairmat-nfdi.github.io/nyaml/tutorials/tutorials.html) is quite cryptic.
