### 2026-04-13 #10231 @maurov
@erickmartins I have almost finished my review. I think that the code and the paper are all fine. Best practices in coding, packaging, and documentation are followed.

Nevertheless, I am asking myself how `nyaml` would really help in practice a user who is not an NeXus expert in writing a definition. It is true that writing YAML is easier than NXDL/XML, but someone who is not familiar with NeXus definitions will not really find a benefit in `nyaml`.

Probably, teaching NeXus is bell beyond `nyaml` scope, so I think the submission could be accepted as it is.
