### 2026-04-16 ec91ac8
add guides for installation, contributing, support
