### 2026-04-16 a2080e0
add page for CLI API
