+++ docs/references/cli-api.md
@@ -1,0 +1,11 @@
+# API for command line tool
+
+`nyaml` supports a single tool for the command line interface (CLI). This page provides documentation for its API.
+
+::: mkdocs-click
+    :module: nyaml.cli
+    :command: launch_tool
+    :prog_name: nyaml2nxdl
+    :depth: 2
+    :style: table
+    :list_subcommands: True
