### 2026-04-16 FAIRmat-NFDI/nyaml#86 @lukaspie
HI @maurov, thanks for starting the review!

I made a fix for this particular issue in #88. We had some small text snippets on contributing scattered across the docs, but I agree it's best to bundle them in one place.

Would you be able to check that this is sufficient? We can always add more information if you think it's helpful. I will have a look at the other issue now as well. Thanks!
