### 2026-04-16 FAIRmat-NFDI/nyaml#87 @lukaspie
Hi @maurov, thanks again for the detailed feedback!

You are right in https://github.com/openjournals/joss-reviews/issues/10231#issuecomment-4237256096 that teaching NeXus from scratch is beyond of what we wanted to do with `nyaml`. The tool is intentionally focused on _just_ transcoding NeXus definitions between YAML and NXDL XML format. Indeed, that means that user require _some_ knowledge for writing NeXus definitions, even when working in YAML. That being said, since `nyaml` also supports the opposite conversion direction (i.e., XML -> YAML), users can at least convert all existing valid NeXus definitions into YAML and start from there for their own schema. Note that `nyaml` is also incorporated as a [make target](https://github.com/nexusformat/definitions/blob/main/Makefile#L38) in the official NIAC repository, so one can even batch convert all existing definitions for reference.

On validation: `nyaml` does not really validate the contents of the YAML file during conversion to XML (or vice versa), with a couple exceptions. For example, it does check for proper nesting (e.g., a field cannot contain a group) and it also emits warning if the `nameType` attribute is used in a "suspicious" way. That being said, standalone validation of NXDL definitions was never in scope for nyaml. The intended workflow is:

- Write or edit a definition in YAML (using nyaml for easier authoring)
- Convert to NXDL XML with nyaml2nxdl
- Add the XML to a fork of the definitions, which runs an [automated validation workflow](https://github.com/nexusformat/definitions/blob/main/.github/workflows/validate.yaml) on every pull request
- And then iterate from there.
