### 2026-04-16 2da2d54
restructure tutorial, add a learning section
