+++ docs/tutorials/tutorials.md
@@ -1,423 +1,0 @@
-# Tutorials for NXDL in YAML format
-This tutorial will explain different keywords, terms, and rules from the perspective of YAML format of the NeXus schema. It provides a overall grasp on how to write a NeXus schema (base classes and application definitions) in YAML format using those syntactical components.
-
-!!! note
-    We do not support the NeXus `choice` concept in the `nyaml` tool yet.
-
-## Design of NeXus Semantics and Terms in YAML
-
-Within the YAML format, the root section denotes the top-level description of the application definition or base class schema, comprising the `category`, `type`, `doc`, `symbols` block, and the name of the schema (e.g. `NXmpes(NXobject)`). Correspondingly, the root section refers to the XML element `definition`, encompassing the first `doc` child of the `definition` and `symbols`. The definition element encapsulates essential XML attributes such as the schema's `name` (an xml attribute), the object it `extends` (an xml attribute), and the schema `type` (an xml attribute), with additional XML attributes (e.g. `xmlns:xsi`) handled by the nyaml converter. The accurate designation of category as either `base` or `application` distinguishes between a `base class` and an `application definition` respectively. The schema name (e.g. `NXmpes(NXobject)`) with parenthesis indicates the extension of the current application definition `NXmpes` from base class `NXobject`, an application definition may extend either `NXobject` or other application definitions. Schemas may incorporate one or multiple symbols, each imbued with specialized physical meanings beyond their literal interpretation, which are utilized over the application definition.
-
-**A typical root section for the application definition `NXmpes` outlined**
-
-=== "YAML"
-    ```yaml
-    category: application
-    type: group
-    doc: |
-      This is the most general application definition for multidimensional photoelectron spectroscopy.
-
-      .. _ISO 18115-1:2023: https://www.iso.org/standard/74811.html
-      .. _IUPAC Recommendations 2020: https://doi.org/10.1515/pac-2019-0404
-    symbols:
-      doc: |
-        The symbols used in the schema to specify e.g. dimensions of arrays
-      n_transmission_function: |
-        Number of data points in the transmission function.
-    NXmpes(NXobject):
-    ```
-
-=== "XML"
-
-    ```xml
-    <definition xmlns="http://definition.nexusformat.org/nxdl/3.1"
-                xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                category="application"
-                type="group"
-                name="NXmpes"
-                extends="NXobject"
-                xsi:schemaLocation="http://definition.nexusformat.org/nxdl/3.1 ../nxdl.xsd">
-      <doc>
-        This is the most general application definition for multidimensional photoelectron spectroscopy.
-        .. _ISO 18115-1:2023: https://www.iso.org/standard/74811.html
-        .. _IUPAC Recommendations 2020: https://doi.org/10.1515/pac-2019-0404
-      </doc>
-      <symbols>
-        <doc>
-          The symbols used in the schema to specify e.g. dimensions of arrays
-        </doc>
-        <symbol name="n_transmission_function">
-          <doc>
-            Number of data points in the transmission function.
-          </doc>
-        </symbol>
-      </symbols>
-    </definition>
-    ```
-### NeXus Group, Field, Attribute, and Link
-#### NeXus Group
[…]
