+++ docs/learn/yaml-notation.md
@@ -1,0 +1,412 @@
+# nyaml notation reference
+
+This page is the detailed reference for writing NeXus definitions in YAML using `nyaml`. It documents all building blocks and special keywords, with YAML and NXDL XML comparisons for each.
+
+For a step-by-step introduction that teaches these concepts progressively through an example, see [Tutorials > Writing a NeXus definition in YAML](../tutorials/writing-a-yaml-definition.md).
+
+!!! note
+    The NeXus `choice` concept is not yet supported in `nyaml`.
+
+---
+
+## Root section
+
+The root section is the top-level block of a definition. It maps to the XML `<definition>` element and contains the following keys:
+
+- `category`: `application` for application definitions, `base` for base classes
+- `type`: always `group`
+- `doc`: free-text description of the definition
+- `symbols`: named dimension constants used throughout the definition (optional)
+- The schema name (e.g. `NXmpes(NXobject)`): the parenthesised base class indicates what this definition extends. An application definition may extend `NXobject` or another application definition.
+
+=== "YAML"
+    ```yaml
+    category: application
+    type: group
+    doc: |
+      This is the most general application definition for multidimensional photoelectron spectroscopy.
+
+      .. _ISO 18115-1:2023: https://www.iso.org/standard/74811.html
+      .. _IUPAC Recommendations 2020: https://doi.org/10.1515/pac-2019-0404
+    symbols:
+      doc: |
+        The symbols used in the schema to specify e.g. dimensions of arrays
+      n_transmission_function: |
+        Number of data points in the transmission function.
+    NXmpes(NXobject):
+    ```
+
+=== "XML"
+
+    ```xml
+    <definition xmlns="http://definition.nexusformat.org/nxdl/3.1"
+                xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                category="application"
+                type="group"
+                name="NXmpes"
+                extends="NXobject"
+                xsi:schemaLocation="http://definition.nexusformat.org/nxdl/3.1 ../nxdl.xsd">
+      <doc>
+        This is the most general application definition for multidimensional photoelectron spectroscopy.
+        .. _ISO 18115-1:2023: https://www.iso.org/standard/74811.html
+        .. _IUPAC Recommendations 2020: https://doi.org/10.1515/pac-2019-0404
+      </doc>
+      <symbols>
+        <doc>
+          The symbols used in the schema to specify e.g. dimensions of arrays
+        </doc>
+        <symbol name="n_transmission_function">
+          <doc>
+            Number of data points in the transmission function.
+          </doc>
+        </symbol>
+      </symbols>
+    </definition>
+    ```
+
+---
+
+## Groups
+
+[NeXus groups](https://manual.nexusformat.org/design.html#design-groups){:target="_blank" rel="noopener"} are instances of NeXus base classes and form the compositional structure of a definition. A group is written as `name(NXbaseclass):` in YAML.
+
+The `nameType` keyword (see below) controls whether the name is matched exactly, loosely, or as a prefix pattern in HDF5 files. Dynamic initialisation (using `nameType: any`) allows multiple instances of the same group type at the same hierarchy level; for example, `(NXmanipulator)` (with `nameType: any`) can be instantiated as `manipulator1` and `manipulator2` during data writing.
+
[…]

[…]
