### 2026-04-16 FAIRmat-NFDI/nyaml#87 @lukaspie
On documentation clarity: I agree the existing tutorial was too dense and assumed too much prior knowledge. We have refactored the documentation in #89 to address this:

- The Writing a NeXus definition in YAML tutorial now has clear prerequisites, a learning-outcome section, and starts with a minimal worked example before going into keyword details

- The How-to guide now includes a section explaining where validation fits in the workflow

- There is a new learning section that explains the YAML notation and how the keywords are supposed to be used.

We also have a separate software tool called `pynxtools` (also currently under review at JOSS), used for converting and validating NeXus file. Since this is typically the first entry point for our users when working with NeXus, we had most of the documentation around NeXus there. In addition to the changes outlined above, I have also added a direct link to the [pynxtools tutorial on writing application definitions](https://fairmat-nfdi.github.io/pynxtools/tutorials/writing-an-application-definition.html), which provides a full end-to-end walkthrough using nyaml.
