### 2026-04-18 #10231 @RubelMozumder
@erickmartins, thanks a lot for your valuable review and feedback.

Someone interested in NeXus to develop their research data model, through the `nyaml` documentation, we always mentione the official website of NeXus https://www.nexusformat.org/ . At the same time, with elaborative examples of different concepts and a full datamodel, the documentation also clearly shows how to write the datamodel in a `yaml` format and convert it to an `xml` file using the `nyaml` converter.

Thanks for your valuable time.
