### 2026-04-21 #10231 @maurov
@erickmartins I have completed my review. Once the related pull request is merged, I would give green light for accepting this contribution in JOSS.
