### 2026-04-21 7d25100
Bump versions for precommit to modernize code linting with the ruff and mypy versions that are typically supported also by typical NOMAD plugins, drop support for python3.9, assure spellchecking for en-US
