+++ .cspell/custom-dictionary.txt
@@ -1,7 +1,4 @@
-# Skip spelling check for the following words
-
-
-# Doc related words
+# Custom Dictionary Words
 mkdocs
 Könnecke
 NXmpes
