+++ .pre-commit-config.yaml
@@ -1,9 +1,29 @@
 repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    # Ruff version.
-    rev: v0.12.0
+    rev: v0.15.11
     hooks:
       # Run the linter.
-      - id: ruff
-      # Run the formatter.
+      - id: ruff-check
+        files: ^(src/nyaml|tests)/.*\.py$
       - id: ruff-format
+        files: ^(src/nyaml|tests)/.*\.py$
+
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v1.20.1
+    hooks:
+      - id: mypy  # static type checking
+
+  - repo: https://github.com/asottile/pyupgrade
+    rev: v3.21.2
+    hooks:
+      - id: pyupgrade
+        args: [--py310-plus]  # modernizes syntax for Python 3.10+
+
+  - repo: https://github.com/streetsidesoftware/cspell-cli
+    rev: v9.8.0
+    hooks:
+      - id: cspell  # spellchecking
+        pass_filenames: false
+        args:
+          - --config
+          - cspell.json
