### 2026-04-21 7ed9490
ran ruff formatting and check and mypy manually to let version bumps take effect
