+++ src/nyaml/nyaml2nxdl.py
@@ -71,7 +71,7 @@ DOM_COMMENT = (
     "# For further information, see http://www.nexusformat.org\n"
 )
 DEPTH_SIZE = 4 * " "
-# Initialised in yml_reader() funtion
+# Initialized in yml_reader() function
 COMMENT_BLOCKS: CommentCollector
 CATEGORY = ""  # Definition would be either 'base' or 'application'
 
@@ -96,7 +96,7 @@ def get_nxdl_copyright_license(nxdl_file):
                         is_comment_end = True
                     else:
                         comment += line
-                # Varifiy for copyright comment
+                # check for copyright comment
                 if is_comment_start and is_comment_end:
                     if is_copyright_comment(comment):
                         return comment
@@ -122,16 +122,16 @@ def set_copyright_text(nxdl_copyright_license=""):
         DOM_COMMENT = DOM_COMMENT.replace("__COPYRIGHT_YEAR__", copyright_year)
 
 
-def yml_reader(inputfile):
+def yml_reader(input_file):
     """
     This function launches the LineLoader class.
     It parses the yaml in a dict and extends it with line tag keys for each key of the dict.
     """
     global COMMENT_BLOCKS
-    with open(inputfile, encoding="utf-8") as plain_text_yaml:
+    with open(input_file, encoding="utf-8") as plain_text_yaml:
         loader = LineLoader(plain_text_yaml)
         loaded_yaml = loader.get_single_data()
-    COMMENT_BLOCKS = CommentCollector(inputfile, loaded_yaml)
+    COMMENT_BLOCKS = CommentCollector(input_file, loaded_yaml)
     COMMENT_BLOCKS.extract_all_comment_blocks()
 
     if "category" not in loaded_yaml.keys():
@@ -166,14 +166,14 @@ def check_for_default_attribute_and_value(xml_element):
     eligible_tag = ["group", "field", "attribute"]
 
     def set_default_attribute(xml_elem, default_attr_to_val):
-        for deflt_attr, deflt_val in default_attr_to_val.items():
+        for default_attribute, default_value in default_attr_to_val.items():
             if (
-                deflt_attr not in xml_elem.attrib
+                default_attribute not in xml_elem.attrib
                 and "maxOccurs" not in xml_elem.attrib
                 and "minOccurs" not in xml_elem.attrib
                 and "recommended" not in xml_elem.attrib
             ):
-                xml_elem.set(deflt_attr, deflt_val)
+                xml_elem.set(default_attribute, default_value)
 
     for child in list(xml_element):
         # skipping comment 'function' that mainly collect comment from yaml file.
@@ -192,11 +192,11 @@ def check_for_default_attribute_and_value(xml_element):
         check_for_default_attribute_and_value(child)
 
 
-def yml_reader_nolinetag(inputfile):
+def yml_reader_no_line_tag(input_file):
     """
     pyyaml based parsing of yaml file in python dict
     """
-    with open(inputfile, encoding="utf-8") as stream:
+    with open(input_file, encoding="utf-8") as stream:
         parsed_yaml = yaml.safe_load(stream)
     return parsed_yaml
 
@@ -262,7 +262,7 @@ def check_for_mapping_char_other(text):
         text = "true"
     if text == "False":
         text = "false"
-    # Some escape char is not valid in yaml libray which is written while writing
+    # Some escape char is not valid in yaml library which is written while writing
     # yaml file. In the time of writing nxdl revert to that escape char.
     escape_reverter = get_yaml_escape_char_reverter_dict()
     for key, val in escape_reverter.items():
@@ -292,11 +292,11 @@ def handle_each_part_doc(text):
     Parameters
     ----------
     text : string
-        String that looks like yaml notaion.
[…]
