+++ src/nyaml/comment_collector.py
@@ -18,8 +18,8 @@
 
 """
 Collect comments in a list by CommentCollector class and each comment is an instance of Comment
-class. Each `Comment` instance(sometimes refered as 'comment block') consists of text and line
-info or neighbours info where the comment must be placed.
+class. Each `Comment` instance(sometimes referred to as 'comment block') consists of text and line
+info or neighboring info where the comment must be placed.
 
 The class Comment is an abstract class for general functions or method to be implemented
 XMLComment and YAMLComment class.
@@ -38,7 +38,7 @@ class CommentCollector:
     _comment_chain.
     """
 
-    def __init__(self, input_file: str = None, loaded_obj: Union[object, dict] = None):
+    def __init__(self, input_file: str = None, loaded_obj: object | dict = None):
         """
         Initialise CommentCollector
         parameters:
@@ -114,20 +114,20 @@ class CommentCollector:
         """
         return self._comment_chain[self._comment_tracker]
 
-    def get_comment_by_line_info(self, comment_locs: tuple[str, Union[int, str]]):
+    def get_comment_by_line_info(self, comment_locs: tuple[str, int | str]):
         """
         Get comment using line information.
         """
         if comment_locs in self._comment_hash:
             return self._comment_hash[comment_locs]
 
-        line_annot, line_loc = comment_locs
-        for cmnt in self._comment_chain:
-            if line_annot in cmnt:
-                line_loc_ = cmnt.get_line_number(line_annot)
+        line_annotation, line_loc = comment_locs
+        for comment in self._comment_chain:
+            if line_annotation in comment:
+                line_loc_ = comment.get_line_number(line_annotation)
                 if line_loc == line_loc_:
-                    self._comment_hash[comment_locs] = cmnt
-                    return cmnt
+                    self._comment_hash[comment_locs] = comment
+                    return comment
 
     def remove_comment(self, ind):
         """Remove a comment from comment list."""
@@ -146,7 +146,7 @@ class CommentCollector:
         """
         Confirm wether the comment corresponds to key_line and line_loc
             is exist or not.
-            comment_locs is equvalant to (line_annotation, line_loc) e.g.
+            comment_locs is equivalent to (line_annotation, line_loc) e.g.
             (__line__doc and 35)
         """
         if not isinstance(comment_locs, tuple):
@@ -154,12 +154,12 @@ class CommentCollector:
                 "Comment_locs should be 'tuple' containing line annotation "
                 "(e.g.__line__doc) and line_loc (e.g. 35)."
             )
-        line_annot, line_loc = comment_locs
-        for cmnt in self._comment_chain:
-            if line_annot in cmnt:
-                line_loc_ = cmnt.get_line_number(line_annot)
+        line_annotation, line_loc = comment_locs
+        for comment in self._comment_chain:
+            if line_annotation in comment:
+                line_loc_ = comment.get_line_number(line_annotation)
                 if line_loc == line_loc_:
-                    self._comment_hash[comment_locs] = cmnt
+                    self._comment_hash[comment_locs] = comment
                     return True
         return False
 
@@ -192,14 +192,14 @@ class Comment:
         """Comment object can be considered as a block element that includes
         document element (an entity for what the comment is written).
         """
-        self._elemt: dict[str, Any] = {}
-        self._elemt_text: str = None
-        self._is_elemt_found: bool = None
-        self._is_elemt_stored: bool = None
+        self._element: dict[str, Any] = {}
+        self._element_text: str = None
+        self._is_element_found: bool = None
+        self._is_element_stored: bool = None
 
-        self._comnt: str = ""
+        self._comment: str = ""
         # If Multiple comments for one element or entity
[…]
