+++ .cspell/custom-dictionary.txt
@@ -10,3 +10,49 @@ nyaml
 NXDL
 lxml
 Forschungsgemeinschaft
+Mozumder
+orcid
+Lukas
+Pielsticker
+Kühbach
+Markus
+Florian
+Dobener
+Sherjeel
+Shabih
+José
+Márquez
+Prieto
+Sandor
+Brockhauser
+Draxl
+Heiko
+zenodo
+utilized
+napimount
+renameable
+nametypes
+xpec
+refindex
+NIAC
+nametype
+AXISNAME
+appdef
+nsmap
+
+
+
+# give better names for these when refactoring (tests)
+someint
+somedocstring
+somevalue
+val
+conceptname
+NXtest
+NXentry
+fileline
+NXfileline
+NXnested
+NXmytests
+NXellipsometry
+NXcomment
