### 2026-04-21 dc4cbba
sort custom dictionary terms lex ascendingly
