+++ .cspell/custom-dictionary.txt
@@ -1,58 +1,55 @@
 # Custom Dictionary Words
-mkdocs
-Könnecke
-NXmpes
-IUPAC
-nxdlformat
-MPES
-infty
-nyaml
-NXDL
-lxml
-Forschungsgemeinschaft
-Mozumder
-orcid
-Lukas
-Pielsticker
-Kühbach
-Markus
-Florian
+AXISNAME
+Brockhauser
 Dobener
-Sherjeel
-Shabih
+Draxl
+Florian
+Forschungsgemeinschaft
+Heiko
+IUPAC
 José
+Könnecke
+Kühbach
+Lukas
+MPES
+Markus
+Mozumder
 Márquez
+NIAC
+NXDL
+NXmpes
+Pielsticker
 Prieto
 Sandor
-Brockhauser
-Draxl
-Heiko
-zenodo
-utilized
-napimount
-renameable
-nametypes
-xpec
-refindex
-NIAC
-nametype
-AXISNAME
+Shabih
+Sherjeel
 appdef
+infty
+lxml
+mkdocs
+nametype
+nametypes
+napimount
 nsmap
-
-
+nxdlformat
+nyaml
+orcid
+refindex
+renameable
+utilized
+xpec
+zenodo
 
 # give better names for these when refactoring (tests)
-someint
+NXcomment
+NXellipsometry
+NXentry
+NXfileline
+NXmytests
+NXnested
+NXtest
+conceptname
 somedocstring
+someint
 somevalue
 val
-conceptname
-NXtest
-NXentry
-fileline
-NXfileline
-NXnested
-NXmytests
-NXellipsometry
-NXcomment
