### 2026-04-21 378e0c0
Fix parsing issue with cff, reintroduced accidentally removed Ref_NXellipsometry
