### 2026-04-21 #10231 @stuartcampbell
@erickmartins I'm happy to recommend acceptance.

The only item to note was that there was no explicit "How to contribute" document or section - but given that the project follows 'normal' development practices - i.e. issues, PRs, etc - then I don't see this as an issue.
