### 2026-04-21 042a0ea
align CI scripts better with those suggested by pynxtools-plugin-template
