+++ .github/workflows/build_docs.yml
@@ -1,34 +1,44 @@
-name: build_and_deploy_docs
+name: Build and deploy mkdocs
+
 on:
   push:
-    branches: [main]
+    branches:
+      - main
+
+env:
+  python-version: 3.12
+
 permissions:
   contents: write
 jobs:
   deploy:
     runs-on: ubuntu-latest
+
     steps:
-      - uses: actions/checkout@v4
-      - name: Configure Git Credentials
+      - name: Checkout repository
+        uses: actions/checkout@v4
+      - name: Configure Git credentials
         run: |
           git config user.name github-actions[bot]
           git config user.email 41898282+github-actions[bot]@users.noreply.github.com
-      - name: Check Spell
-        uses: streetsidesoftware/cspell-action@v7
-        with:
-          files: "docs/**/*.md"
-          config: .cspell/cspell.config.yaml
-          strict: true
 
-      - uses: actions/setup-python@v4
+      - name: Install uv and set the python version to ${{ env.python-version }}
+        uses: astral-sh/setup-uv@v5
         with:
-          python-version: 3.x
-      - run: echo "cache_id=$(date --utc '+%V')" >> $GITHUB_ENV
-      - uses: actions/cache@v3
+          python-version: ${{ env.python-version }}
+
+      - name: Cache mkdocs-material environment
+        uses: actions/cache@v3
         with:
           key: mkdocs-material-${{ env.cache_id }}
           path: .cache
           restore-keys: |
             mkdocs-material-
-      - run: pip install ".[docs]"
-      - run: mkdocs gh-deploy --force --remote-branch gh-pages
+
+      - name: Install dependencies
+        run: |
+          uv pip install ".[docs]"
+
+      - name: Build and deploy
+        run: |
+          mkdocs gh-deploy --force --remote-branch gh-pages
