+++ .github/workflows/pylint.yaml
@@ -2,29 +2,24 @@ name: linting
 
 on: [push]
 
+env:
+  python-version: 3.12
 jobs:
   linting:
     runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
-
     steps:
       - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
       - name: Install uv and set the python version to ${{ matrix.python-version }}
         uses: astral-sh/setup-uv@v5
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
+      - name: Install package and dev dependencies
         run: |
           uv pip install ".[dev]"
       - name: ruff check
         run: |
           ruff check src/nyaml tests
-      - name: ruff formatting
+      - name: ruff format
         run: |
           ruff format --check src/nyaml tests
       - name: mypy
