+++ .github/workflows/spellcheck.yml
@@ -1,0 +1,35 @@
+name: Spell check
+
+on:
+  pull_request:
+    branches: [main]
+
+jobs:
+  spellcheck:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v4
+
+    - name: Install dependencies and dictionary
+      run: |
+        npm install -g cspell jq \
+        @cspell/dict-en-us \
+        @cspell/dict-python \
+        @cspell/dict-scientific-terms-us
+
+    - name: Run cspell
+      run: |
+        cspell --config cspell.json \
+        "README.md" "CITATION.cff" \
+        "docs/**/*.md" \
+        "src/nyaml/**/*.py" \
+        "tests/**/*.py"
+
+    - name: Fail on errors
+      run: |
+        if [[ $? -ne 0 ]]; then
+          echo "There are spelling errors in your code. Please fix them and try again."
+          exit 1
+        fi
