+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
       - id: cspell  # spellchecking
         pass_filenames: false
         args:
-          - --unique
+          # - --unique
           - --config
           - cspell.json
           - README.md
