### 2026-04-21 94cb55e
fix issues with CI scripts
