### 2026-04-21 5da05af
update publish CI script
