### 2026-04-27 59177bb
Apply suggestions from code review
Co-authored-by: Markus Kühbach <mkuehbach@users.noreply.github.com>
Co-authored-by: RubelMozumder <32923026+RubelMozumder@users.noreply.github.com>
