### 2026-04-27 29465dc
add guides for installation, contributing, support (#88)
Co-authored-by: lukaspie <lukaspie@github.com>
