+++ docs/contact.md
@@ -1,0 +1,16 @@
+# Get in contact
+
+`nyaml` is an open source project that warmly welcomes community projects, contributions, suggestions, bug fixes, and constructive feedback. `nyaml` is built mainly within FAIRmat Area B - Experiment.
+
+You can reach us through different channels. You can send an email directly to the main contributors list:
+
+| Name | E-mail     | GitHub profile |
+|------|------------|-----------------|
+| Rubel Mozumder | [rubel.mozumder@physik.hu-berlin.de](mailto:rubel.mozumder@physik.hu-berlin.de) | [@RubelMozumder](https://github.com/RubelMozumder) |
+| Dr. Markus Kühbach | [markus.kuehbach@physik.hu-berlin.de](mailto:markus.kuehbach@physik.hu-berlin.de) | [@mkuehbach](https://github.com/mkuehbach) |
+| Dr. Lukas Pielsticker | [lukas.pielsticker@physik.hu-berlin.de](mailto:lukas.pielsticker@physik.hu-berlin.de) |  [@lukaspie](https://github.com/lukaspie) |
+
+Alternatively, you can also:
+
+- Open an issue on the [nyaml GitHub](https://github.com/FAIRmat-NFDI/nyaml/issues)
+- Join the [NOMAD Discord channel](https://discord.gg/Gyzx3ukUw8) and ask us there directly.
