### 2026-04-27 54d8cb8
restructure tutorial, add a learning section (#89)
Co-authored-by: Markus Kühbach <mkuehbach@users.noreply.github.com>
Co-authored-by: RubelMozumder <32923026+RubelMozumder@users.noreply.github.com>
Co-authored-by: lukaspie <lukaspie@github.com>
