+++ docs/index.md
@@ -1,27 +1,32 @@
 # nyaml documentation
 
-The `nyaml` tool offers a user-friendly command-line interface for converting between XML and YAML representations of NeXus definitions written in the NeXus Definition Language (NXDL). While the official NeXus definitions are maintained in `XML,` authoring them in YAML provides a more concise and readable alternative. `nyaml` bridges this gap, allowing developers to work in YAML while ensuring compatibility with the official XML format.
+The `nyaml` tool offers a user-friendly command-line interface for converting between XML and YAML representations of NeXus definitions written in the NeXus Definition Language (NXDL). While the official NeXus definitions are maintained in XML, authoring them in YAML provides a more concise and readable alternative. `nyaml` bridges this gap, allowing developers to work in YAML while ensuring compatibility with the official XML format.
 
 The tool supports both NeXus application definitions and base classes. It can be installed via `pip` and includes comprehensive documentation and tutorials to assist users in writing NeXus definitions using YAML.
 
 <div markdown="block" class="home-grid">
 <div markdown="block">
 
-### Tutorial
+### Tutorials
 
-Tutorials to write different parts and a full NeXus application or base class
+Step-by-step guides for installing `nyaml` and writing NeXus definitions in YAML
 
 - [Installation](tutorials/installation.md)
+<<<<<<< HEAD
 - [Writing a NeXus definition in YAML](tutorials/tutorials.md)
 - [Contributing](tutorials/contributing.md)
 
+=======
+- [Writing a NeXus definition in YAML](tutorials/writing-a-yaml-definition.md)
+- [Contributing](tutorials/contributing.md)
+>>>>>>> 2262a53 (restructure tutorial, add a learning section)
 
 </div>
 <div markdown="block">
 
 ### How-to guides
 
-How-to install and use the nyaml tool
+How to install and use the `nyaml` tool
 
 - [Getting started with `nyaml`](howtos/howto.md)
 
@@ -33,19 +38,25 @@ How-to install and use the nyaml tool
 
 An introduction to NeXus and its use in the light of YAML format using the `nyaml` tool.
 
-- [An introduction to NeXus](https://manual.nexusformat.org/index.html)
-- [Motivation behind the nyaml tool](learn/explanations.md)
+- [Introduction to NeXus](https://manual.nexusformat.org/index.html){:target="_blank" rel="noopener"}
+- [Motivation behind the `nyaml` tool](learn/explanations.md)
 - [Workflow of the nyaml tool with examples](learn/workflow.md)
-
+- [nyaml notation reference](learn/yaml-notation.md)
 
 </div>
 <div markdown="block">
 
 ### References
+
 The list of references used in this documentation
 
+<<<<<<< HEAD
 - [Command line API](references/cli-api.md)
+=======
+- [CLI API](references/cli-api.md)
+>>>>>>> 2262a53 (restructure tutorial, add a learning section)
 - [References](references/references.md)
+
 </div>
 
 </div>
@@ -54,10 +65,10 @@ The list of references used in this documentation
 
 For questions or suggestions:
 
-- Open an issue on the [`nyaml` GitHub](https://github.com/FAIRmat-NFDI/nyaml/issues)
-- Join our [Discord channel](https://discord.gg/Gyzx3ukUw8)
+- Open an issue on the [`nyaml` GitHub](https://github.com/FAIRmat-NFDI/nyaml/issues){:target="_blank" rel="noopener"}
+- Join our [Discord channel](https://discord.gg/Gyzx3ukUw8){:target="_blank" rel="noopener"}
 - Get in contact with our [lead developers](contact.md).
 
 <h2>Project and community</h2>
 
[…]
