+++ src/nyaml/nyaml2nxdl.py
@@ -27,6 +27,7 @@ import re
 import textwrap
 from typing import Any
 from urllib.parse import unquote
+import warnings
 
 import lxml.etree as ET
 import yaml
@@ -215,10 +216,10 @@ def check_for_skipped_attributes(
     Check for any attributes have been skipped or not.
     NOTE: We should keep in mind about 'doc'
     """
-    block_tag = ["enumeration"]
+    block_tag = [r"\enumeration"]
     if value:
         for attr, val in value.items():
-            if attr == "doc":
+            if attr == r"\doc":
                 continue
             if "__line__" in attr or attr in block_tag:
                 continue
@@ -227,7 +228,7 @@ def check_for_skipped_attributes(
                 print(f"__line__ : {value[line_number]}")
             if (
                 not isinstance(val, dict)
-                and "\\@" not in attr
+                and r"\@" not in attr
                 and attr not in allowed_attr
                 and "NX" not in attr
                 and val
@@ -435,6 +436,26 @@ def xml_handle_dimensions(
     line_loc = dct[line_number]
     dims: ET.Element | None = None
     if isinstance(value, dict):
+        # Normalise escape-prefixed keywords inside a dimensions block to their bare
+        # internal names (which the rest of this function uses). Also accept the old
+        # bare forms with a DeprecationWarning where applicable.
+        value = dict(value)  # shallow copy so we do not mutate the caller's dict
+        for _kw in (r"\rank", r"\doc", r"\dim"):
+            bare = _kw[1:]  # strip leading backslash
+            if _kw in value:
+                value[bare] = value.pop(_kw)
+                line_key = f"__line__{_kw}"
+                if line_key in value:
+                    value[f"__line__{bare}"] = value.pop(line_key)
+        if "rank" in value and r"\rank" not in value:
+            # Accept old bare 'rank' with a deprecation warning
+            warnings.warn(
+                r"Use '\rank' instead of 'rank' as the key inside a 'dimensions' block. "
+                "Support for the unescaped 'rank' keyword is deprecated and will be "
+                "removed in a future version.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
         # top-level docstring dealt with already by the caller
         n_idx_dicts = len(
             [key for key in value if re.match("^[0-9]+$", f"{key}") is not None]
@@ -611,8 +632,6 @@ def xml_handle_link(
     link_obj = ET.SubElement(obj, "link")
     link_obj.set("name", str(name))
 
-    xml_handle_nametype(keyword, keyword, dct, obj)
-
     if value:
         rm_key_list = []
         for attr, val in value.items():
@@ -620,8 +639,17 @@ def xml_handle_link(
                 continue
             line_number = f"__line__{attr}"
             line_loc = value[line_number]
-            if attr == "doc":
+            if attr == r"\doc":
                 xml_handle_doc(link_obj, val, line_number, line_loc)
+                rm_key_list.append(attr)
+                rm_key_list.append(line_number)
+            elif attr == r"\nameType":
+                link_name = keyword[:-6]  # strip "(link)"
+                check_for_proper_nameType(
+                    link_name, str(val) if val else None, keyword
+                )
+                if val:
+                    link_obj.set("nameType", str(val))
                 rm_key_list.append(attr)
                 rm_key_list.append(line_number)
             elif attr in YAML_LINK_ATTRIBUTES and not isinstance(val, dict):
@@ -663,7 +691,7 @@ def xml_handle_choice(
                 continue
             line_number = f"__line__{attr}"
             line_loc = value[line_number]
-            if attr == "doc":
+            if attr == r"\doc":
[…]
