+++ src/nyaml/nxdl2nyaml.py
@@ -35,6 +35,7 @@ from nyaml.helper import (
     NXDL_FIELD_ATTRIBUTES,
     NXDL_GROUP_ATTRIBUTES,
     NXDL_LINK_ATTRIBUTES,
+    RESERVED_KEYWORDS,
     check_for_proper_nameType,
     clean_empty_lines,
     get_node_parent_info,
@@ -407,6 +408,8 @@ class Nxdl2yaml:
         """
         if "}" in tag:
             tag = remove_namespace_from_tag(tag)
+        if depth > 0 and tag in RESERVED_KEYWORDS:
+            tag = f"\\{tag}"
         indent = depth * DEPTH_SIZE
         text = self.clean_and_organize_text(text, depth)  # starts with '\n'
         docs = re.split(r"\n\s*\n", text)
@@ -636,24 +639,26 @@ class Nxdl2yaml:
             self.check_for_unwanted_attributes(node=node)
             # As both 'minOccurs', 'maxOccurs' and optionality move to the 'exists'
             if key in self.optionality_keys:
-                if "exists" not in tmp_dict:
-                    tmp_dict["exists"] = []
+                if r"\exists" not in tmp_dict:
+                    tmp_dict[r"\exists"] = []
                 self.handle_exists(exists_dict, key, val)
             elif key == "units":
-                tmp_dict["unit"] = str(val)
+                tmp_dict[r"\unit"] = str(val)
             else:
-                tmp_dict[key] = str(val)
+                # Escape reserved keywords so they are not misinterpreted on round-trip
+                escaped_key = f"\\{key}" if key in RESERVED_KEYWORDS else key
+                tmp_dict[escaped_key] = str(val)
 
         if exists_dict:
             for key, exists_val in exists_dict.items():
                 if key in ["minOccurs", "maxOccurs"]:
-                    tmp_dict["exists"] = (
-                        tmp_dict["exists"] + exists_val
+                    tmp_dict[r"\exists"] = (
+                        tmp_dict[r"\exists"] + exists_val
                         if isinstance(exists_val, list)
                         else [exists_val]
                     )
                 elif key in ["optional", "recommended", "required"]:
-                    tmp_dict["exists"] = key
+                    tmp_dict[r"\exists"] = key
 
         depth_ = depth + 1
         for key, val_ in tmp_dict.items():
@@ -710,7 +715,7 @@ class Nxdl2yaml:
             if not isinstance(val, dict):
                 if attr in ["rank"]:
                     # indent = (depth + 1) * DEPTH_SIZE
-                    yml_dim_dct["rank"] = val
+                    yml_dim_dct[r"\rank"] = val
                     break
                 # rank is the only allowed attribute of a dimensionsType node
                 # see https://manual.nexusformat.org/nxdl_desc.html#dimensionstype
@@ -759,9 +764,9 @@ class Nxdl2yaml:
         # perform I/O based on the cases analyzed
         yml_dim_dct_keys = list(yml_dim_dct)
         indent = depth * DEPTH_SIZE
-        if set(yml_dim_dct_keys) in [{"rank"}, {"doc", "rank"}]:
+        if set(yml_dim_dct_keys) in [{r"\rank"}, {"doc", r"\rank"}]:
             # rank only notation
-            file_out.write(f"{indent}dimensions:\n")
+            file_out.write(f"{indent}\\dimensions:\n")
             for key, val in yml_dim_dct.items():
                 if key == "doc":
                     file_out.write(f"{val}")
@@ -776,10 +781,10 @@ class Nxdl2yaml:
                             use_shorthand_notation = False
                             break
             if use_shorthand_notation:  # shorthand_explicit_rank_new
[…]
