+++ .cspell/custom-dictionary.txt
@@ -11,6 +11,7 @@ José
 Könnecke
 Kühbach
 Lukas
+lukaspie
 MPES
 Markus
 Mozumder
@@ -24,17 +25,19 @@ Sandor
 Shabih
 Sherjeel
 appdef
+fspath
 infty
 lxml
 mkdocs
+mkuehbach
 nametype
 nametypes
 napimount
 nsmap
 nxdlformat
 nyaml
-pynxtools
 orcid
+pynxtools
 refindex
 renameable
 utilized
