+++ .github/workflows/spellcheck.yml
@@ -20,16 +20,19 @@ jobs:
         @cspell/dict-scientific-terms-us
 
     - name: Run cspell
+      id: cspell
+      continue-on-error: true
       run: |
         cspell --config cspell.json \
         "README.md" "CITATION.cff" \
         "docs/**/*.md" \
         "src/nyaml/**/*.py" \
-        "tests/**/*.py"
+        "tests/**/*.py" 2>&1 | tee cspell.log
 
     - name: Fail on errors
+      if: steps.cspell.outcome == 'failure'
       run: |
-        if [[ $? -ne 0 ]]; then
-          echo "There are spelling errors in your code. Please fix them and try again."
-          exit 1
-        fi
+        echo "There are spelling errors in your code. Please fix them and try again."
+        echo "++++++++++ cspell error ++++++++++"
+        cat cspell.log
+        exit 1
