+++ .github/workflows/build_docs.yml
@@ -27,10 +27,14 @@ jobs:
         with:
           python-version: ${{ env.python-version }}
 
+      - name: Set cache key suffix
+        id: cache_key
+        run: echo "value=$(date --utc '+%V')" >> "$GITHUB_OUTPUT"
+
       - name: Cache mkdocs-material environment
         uses: actions/cache@v3
         with:
-          key: mkdocs-material-${{ env.cache_id }}
+          key: mkdocs-material-${{ steps.cache_key.outputs.value }} # Restore content with the same cache key
           path: .cache
           restore-keys: |
             mkdocs-material-
