### 2026-04-28 #10231 @RubelMozumder
> [@erickmartins](https://github.com/erickmartins) I'm happy to recommend acceptance.
>
> The only item to note was that there was no explicit "How to contribute" document or section - but given that the project follows 'normal' development practices - i.e. issues, PRs, etc - then I don't see this as an issue.

Hi @stuartcampbell,
Thank you for your positive feedback. We appreciate your time and effort to review this paper.
We have included a section on how to contribute to the code repository. We have also refactored the docs about how to write NeXus definition in `yaml` format with a detailed explanation of `yaml` syntax in this regard.

Relavent PRs (merged):
[PR-89](https://github.com/FAIRmat-NFDI/nyaml/pull/89)
[PR-88](https://github.com/FAIRmat-NFDI/nyaml/pull/88)
