+++ src/nyaml/helper.py
@@ -24,7 +24,9 @@ to convert from nyaml to nxdl and vice versa.
 """
 
 import hashlib
+import os
 from collections.abc import Callable
+from typing import Any
 
 from yaml.composer import Composer
 from yaml.constructor import Constructor
@@ -33,11 +35,13 @@ from yaml.nodes import ScalarNode
 from yaml.resolver import BaseResolver
 
 # Yaml library does not except the keys (escape char "\t" and yaml separator ":")
-ESCAPE_CHAR_DICT_IN_YAML = {"\t": "    "}
-ESCAPE_CHAR_DICT_IN_XML = {val: key for key, val in ESCAPE_CHAR_DICT_IN_YAML.items()}
+ESCAPE_CHAR_DICT_IN_YAML: dict[str, str] = {"\t": "    "}
+ESCAPE_CHAR_DICT_IN_XML: dict[str, str] = {
+    val: key for key, val in ESCAPE_CHAR_DICT_IN_YAML.items()
+}
 
 # Set up attributes for nxdl version
-NXDL_GROUP_ATTRIBUTES = (
+NXDL_GROUP_ATTRIBUTES: tuple[str, ...] = (
     "optional",
     "recommended",
     "name",
@@ -47,7 +51,7 @@ NXDL_GROUP_ATTRIBUTES = (
     "deprecated",
     "nameType",
 )
-NXDL_FIELD_ATTRIBUTES = (
+NXDL_FIELD_ATTRIBUTES: tuple[str, ...] = (
     "optional",
     "recommended",
     "name",
@@ -68,7 +72,7 @@ NXDL_FIELD_ATTRIBUTES = (
     "units",
 )
 
-NXDL_ATTRIBUTES_ATTRIBUTES = (
+NXDL_ATTRIBUTES_ATTRIBUTES: tuple[str, ...] = (
     "name",
     "type",
     "recommended",
@@ -77,31 +81,42 @@ NXDL_ATTRIBUTES_ATTRIBUTES = (
     "nameType",
 )
 
-NXDL_LINK_ATTRIBUTES = ("name", "target", "napimount", "nameType")
+NXDL_LINK_ATTRIBUTES: tuple[str, ...] = ("name", "target", "napimount", "nameType")
 
 # Set up attributes for yaml version
-YAML_GROUP_ATTRIBUTES = (*NXDL_GROUP_ATTRIBUTES, "exists")
+YAML_GROUP_ATTRIBUTES: tuple[str, ...] = (*NXDL_GROUP_ATTRIBUTES, "exists")
 
-YAML_FIELD_ATTRIBUTES = (*NXDL_FIELD_ATTRIBUTES[0:-1], "unit", "exists", "dim")
+YAML_FIELD_ATTRIBUTES: tuple[str, ...] = (
+    *NXDL_FIELD_ATTRIBUTES[0:-1],
+    "unit",
+    "exists",
+    "dim",
+)
 
-YAML_ATTRIBUTES_ATTRIBUTES = (
+YAML_ATTRIBUTES_ATTRIBUTES: tuple[str, ...] = (
     *NXDL_ATTRIBUTES_ATTRIBUTES,
     "minOccurs",
     "maxOccurs",
     "exists",
 )
 
-YAML_LINK_ATTRIBUTES = NXDL_LINK_ATTRIBUTES
+YAML_LINK_ATTRIBUTES: tuple[str, ...] = NXDL_LINK_ATTRIBUTES
 
 
-def remove_namespace_from_tag(tag):
+def remove_namespace_from_tag(tag: str | Callable[..., Any]) -> str:
     """Helper function to remove the namespace from an XML tag."""
-    if isinstance(tag, Callable) and tag.__name__ == "Comment":
[…]

[…]
