### 2026-04-28 ea69837
use local env for mypy.
