### 2026-04-28 ac753cf
split into yaml2nxdl, nxdl2yaml, and roundtrip tests; move test files accordingly
