### 2026-04-29 8e45d8c
Remove the lxml-stubs from dev that take precedence on types-lxml and treat xml element as HTML element.
