### 2026-04-29 82dee88
Modernizing linting, code styling, and supported py versions (#90)
* Bump versions for precommit to modernize code linting with the ruff and mypy versions that are typically supported also by typical NOMAD plugins, drop support for python3.9, assure spellchecking for en-US

* ran ruff formatting and check and mypy manually to let version bumps take effect

* sort custom dictionary terms lex ascendingly

* Fix parsing issue with cff, reintroduced accidentally removed Ref_NXellipsometry

* align CI scripts better with those suggested by pynxtools-plugin-template

* fix issues with CI scripts

* 
[…]
