+++ src/nyaml/comment_collector.py
@@ -18,14 +18,16 @@
 
 """
 Collect comments in a list by CommentCollector class and each comment is an instance of Comment
-class. Each `Comment` instance(sometimes refered as 'comment block') consists of text and line
-info or neighbours info where the comment must be placed.
+class. Each `Comment` instance(sometimes referred to as 'comment block') consists of text and line
+info or neighboring info where the comment must be placed.
 
 The class Comment is an abstract class for general functions or method to be implemented
 XMLComment and YAMLComment class.
 """
 
-from typing import Any, Union
+import os
+from collections.abc import Iterator
+from typing import Any
 
 from nyaml.helper import LineLoader
 
@@ -38,18 +40,22 @@ class CommentCollector:
     _comment_chain.
     """
 
-    def __init__(self, input_file: str = None, loaded_obj: Union[object, dict] = None):
+    def __init__(
+        self,
+        input_file: str | os.PathLike[str] = None,
+        loaded_obj: object | dict = None,
+    ):
         """
         Initialise CommentCollector
         parameters:
             input_file: raw input file (xml, yml)
             loaded_obj: file loaded by third party library
         """
-        self._comment_chain: list = []
-        self.file = input_file
+        self._comment_chain: list[YAMLComment | XMLComment] = []
+        self.file: str = os.fspath(input_file)
         self._comment_tracker = 0
-        self._comment_hash: dict[tuple, type[Comment]] = {}
-        self.comment: type[Comment]
+        self._comment_hash: dict[tuple, Comment] = {}
+        self.comment: type[XMLComment] | type[YAMLComment] = None
         if self.file and not loaded_obj:
             if self.file.endswith(".xml"):
                 self.comment = XMLComment
@@ -72,7 +78,7 @@ class CommentCollector:
         else:
             raise ValueError("Incorrect inputs for CommentCollector")
 
-    def extract_all_comment_blocks(self):
+    def extract_all_comment_blocks(self) -> None:
         """Collect all comments.
 
         Note here that comment means (comment text + element or line info
@@ -108,45 +114,48 @@ class CommentCollector:
                     single_comment = self.comment(last_comment=single_comment)
                     single_comment.process_each_line(line, (line_num + 1))
 
-    def get_next_comment(self):
+    def get_next_comment(self) -> "YAMLComment | XMLComment":
         """
         Return comment from comment_chain that must come earlier in order.
         """
         return self._comment_chain[self._comment_tracker]
 
-    def get_comment_by_line_info(self, comment_locs: tuple[str, Union[int, str]]):
+    def get_comment_by_line_info(
+        self, comment_locs: tuple[str, int | str]
+    ) -> "Comment | None":
         """
         Get comment using line information.
         """
         if comment_locs in self._comment_hash:
             return self._comment_hash[comment_locs]
 
-        line_annot, line_loc = comment_locs
-        for cmnt in self._comment_chain:
-            if line_annot in cmnt:
-                line_loc_ = cmnt.get_line_number(line_annot)
+        line_annotation, line_loc = comment_locs
+        for comment in self._comment_chain:
+            if isinstance(comment, YAMLComment) and line_annotation in comment:
+                line_loc_ = comment.get_line_number(line_annotation)
                 if line_loc == line_loc_:
-                    self._comment_hash[comment_locs] = cmnt
-                    return cmnt
+                    self._comment_hash[comment_locs] = comment
+                    return comment
+        return None
 
-    def remove_comment(self, ind):
+    def remove_comment(self, ind: int) -> None:
         """Remove a comment from comment list."""
         if ind < len(self._comment_chain):
             del self._comment_chain[ind]
         else:
             raise ValueError("Oops! Index is out of range.")
 
[…]
