+++ src/nyaml/nxdl2nyaml.py
@@ -22,10 +22,11 @@ This file collects the functions for conversion from nxdl.xml to yaml version.
 
 # pylint: disable=too-many-lines
 
+import os
 import re
 import textwrap
 from pathlib import Path
-from typing import Callable
+from typing import Any, TextIO
 
 import lxml.etree as ET
 
@@ -43,36 +44,36 @@ from nyaml.helper import (
 )
 
 DEPTH_SIZE = 2 * " "
-CMNT_TAG = "!--"
-CMNT_TAG_END = "--"
-CMNT_START = "<!--"
-CMNT_END = "-->"
+COMMENT_TAG = "!--"
+COMMENT_TAG_END = "--"
+COMMENT_START = "<!--"
+COMMENT_END = "-->"
 DEFINITION_CATEGORIES = ("category: application", "category: base")
 
 
-def separate_pi_comments(input_file):
+def separate_pi_comments(input_file: str | os.PathLike[str]) -> list[str]:
     """Separate PI comments from ProcessesInstruction (PI).
 
     ProcessesInstruction refers xml element process and version
     Separate the comments that comes immediately after XML process instruction part,
     i.e. copyright comment part.
     """
-    comments_list = []
-    comment = []
+    comments_list: list[str] = []
+    comment: list[str] = []
 
     with open(input_file, encoding="utf-8") as file:
         lines = file.readlines()
         def_tag = "<definition"
         for line in lines:
-            if CMNT_START in line:
-                line = line.replace(CMNT_START, "")
-                if CMNT_END in line:
-                    line = line.replace(CMNT_END, "")
+            if COMMENT_START in line:
+                line = line.replace(COMMENT_START, "")
+                if COMMENT_END in line:
+                    line = line.replace(COMMENT_END, "")
                     comments_list.append(line)
                 else:
                     comment.append(line)
-            elif CMNT_END in line and len(comment) > 0:
-                comment.append(line.replace(CMNT_END, ""))
+            elif COMMENT_END in line and len(comment) > 0:
+                comment.append(line.replace(COMMENT_END, ""))
                 comments_list.append("".join(comment))
                 comment.clear()
             elif len(comment) > 0:
@@ -89,12 +90,12 @@ class _CommentedTreeBuilder(ET.TreeBuilder):
 
     def Comment(self, text):  # pylint: disable=invalid-name
         """Defining comment builder in TreeBuilder"""
-        self.start(CMNT_TAG, {})
+        self.start(COMMENT_TAG, {})
         self.data(text)
-        self.end(CMNT_TAG_END)
+        self.end(COMMENT_TAG_END)
 
 
-def parse(filepath):
+def parse(filepath: str | os.PathLike[str]) -> tuple[list[str], ET._ElementTree]:
     """Parse xml function.
 
     Construct parser function for modified tree builder for including modified TreeBuilder
@@ -107,7 +108,9 @@ def parse(filepath):
     return comments, root
 
 
-def handle_mapping_char(text, depth=-1, skip_n_line_on_top=False):
+def handle_mapping_char(
+    text: str, depth: int = -1, skip_n_line_on_top: bool = False
+) -> str:
     """Check for escape character and replace by alternative character."""
 
     escape_char = get_yaml_escape_char_dict()
@@ -122,7 +125,7 @@ def handle_mapping_char(text, depth=-1, skip_n_line_on_top=False):
     return text
 
 
-def add_new_line_with_pipe_on_top(text, depth):
+def add_new_line_with_pipe_on_top(text: str, depth: int) -> str:
     """Design docs block.
 
     Restructure the text by adding pipe '|' and '\n' char if ':' is found in text.
@@ -141,18 +144,18 @@ class Nxdl2yaml:
         self,
[…]
