+++ .cspell/custom-dictionary.txt
@@ -1,15 +1,59 @@
-# Skip spelling check for the following words
-
-
-# Doc related words
-mkdocs
-Könnecke
-NXmpes
-IUPAC
-nxdlformat
-MPES
-infty
-nyaml
-NXDL
-lxml
+# Custom Dictionary Words
+AXISNAME
+Brockhauser
+Dobener
+Draxl
+Florian
 Forschungsgemeinschaft
+Heiko
+IUPAC
+José
+Könnecke
+Kühbach
+Lukas
+lukaspie
+MPES
+Markus
+Mozumder
+Márquez
+NIAC
+NXDL
+NXmpes
+Pielsticker
+Prieto
+Sandor
+Shabih
+Sherjeel
+appdef
+fspath
+infty
+lxml
+mkdocs
+mkuehbach
+nametype
+nametypes
+napimount
+nsmap
+nxdlformat
+nyaml
+orcid
+pynxtools
+refindex
+renameable
+utilized
+xpec
+zenodo
+
+# give better names for these when refactoring (tests)
+NXcomment
+NXellipsometry
+NXentry
+NXfileline
+NXmytests
+NXnested
+NXtest
+conceptname
+somedocstring
+someint
+somevalue
+val
