+++ tests/helpers.py
@@ -1,0 +1,61 @@
+#
+# Copyright The NOMAD Authors.
+#
+# This file is part of NOMAD. See https://nomad-lab.eu for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""Shared test helper utilities for nyaml tests."""
+
+import re
+from datetime import datetime
+from pathlib import Path
+
+LATEST_COPYRIGHT_YEAR = f"{datetime.now().year}-{datetime.now().year}"
+LATEST_COPYRIGHT = rf"# Copyright \(C\) {LATEST_COPYRIGHT_YEAR} NeXus International Advisory Committee \(NIAC\)"
+COPYRIGHT_REPLACEMENT = (
+    "# Copyright (C) 2010-2020 NeXus International Advisory Committee (NIAC)"
+)
+
+
+def check_and_replace_latest_copyright(nxdl_file):
+    """Check if the latest copyright date has been written properly,
+    and replace it accordingly to the test data file.
+    """
+    content = nxdl_file.read_text()
+    generated_copyright = re.findall(LATEST_COPYRIGHT, content, re.DOTALL)
+    assert len(generated_copyright) == 1, (
+        f"No copyright or not correct copyright year found in {nxdl_file}"
+    )
+    content = re.sub(LATEST_COPYRIGHT, COPYRIGHT_REPLACEMENT, content)
+    nxdl_file.write_text(content)
+
+
+def check_file_fresh_baked(test_file):
+    """Get sure that the test file is generated by the converter."""
+    path = Path(test_file)
+    timestamp = datetime.fromtimestamp(path.stat().st_mtime).strftime("%d/%m/%Y %H:%M")
+    now = datetime.now().strftime("%d/%m/%Y %H:%M")
+    assert timestamp == now, "xml file not generated"
+
+
+def compare_yaml_content(yaml_dict1, yaml_dict2, test_key_list):
+    """Recursively compare two YAML dicts for a given list of keys."""
+    for k_val1, k_val2 in zip(yaml_dict1.items(), yaml_dict2.items()):
+        key1, val1 = k_val1
+        key2, val2 = k_val2
+        for ref_key in test_key_list:
+            if key1 == ref_key and key2 == ref_key:
+                assert val1 == val2, f"{ref_key} content is not the same."
+            elif isinstance(val1, dict) and isinstance(val2, dict):
+                compare_yaml_content(val1, val2, test_key_list)
