+++ tests/test_nxdl2nyaml.py
@@ -1,0 +1,165 @@
+#
+# Copyright The NOMAD Authors.
+#
+# This file is part of NOMAD. See https://nomad-lab.eu for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""
+Tests for the NXDL → YAML conversion direction (nxdl2nyaml).
+"""
+
+import filecmp
+import os
+import sys
+from pathlib import Path
+
+import pytest
+from click.testing import CliRunner
+from helpers import check_file_fresh_baked, compare_yaml_content
+
+from nyaml import cli as nyaml2nxdl
+from nyaml.helper import LineLoader
+
+
+def test_nxdl2yaml_nameType():
+    """
+    Check the correct handling of the nameType attribute for the direction
+    nxdl->nyaml.
+    """
+    ref_xml_file = "tests/data/nxdl2yaml/allowed_nameType.nxdl.xml"
+    ref_yml_file = "tests/data/nxdl2yaml/ref_allowed_nameType.yaml"
+    test_yml_file = "tests/data/nxdl2yaml/allowed_nameType_parsed.yaml"
+
+    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
+    assert result.exit_code == 0
+    check_file_fresh_baked(test_yml_file)
+
+    result = filecmp.cmp(ref_yml_file, test_yml_file, shallow=False)
+
+    assert result, "Ref YML and parsed YML don't have the same structure!"
+    os.remove(test_yml_file)
+    sys.stdout.write("Test on xml -> yml nameType okay.\n")
+
+    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
+    assert result.exit_code == 0
+    check_file_fresh_baked(test_yml_file)
+
+    result = filecmp.cmp(ref_yml_file, test_yml_file, shallow=False)
+    assert result, (
+        "Ref YML and parsed YML\
+has not the same structure!!"
+    )
+    os.remove(test_yml_file)
+    sys.stdout.write("Test on xml -> yml doc formatting okay.\n")
+
+
+def test_nxdl2yaml_doc_format_and_nxdl_part_as_comment():
+    """
+    This test for two reasons:
+            1. In test-1 an nxdl file with all kind of doc formats are translated
+    to yaml to check if they are correct.
+            2. In test-2: Check the nxdl that comes at the end of yaml file as comment.
+    """
+    ref_xml_file = "tests/data/Ref_NXentry.nxdl.xml"
+    ref_yml_file = "tests/data/Ref_NXentry.yaml"
+    test_yml_file = "tests/data/Ref_NXentry_parsed.yaml"
+    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
+    assert result.exit_code == 0
+    check_file_fresh_baked(test_yml_file)
+
+    result = filecmp.cmp(ref_yml_file, test_yml_file, shallow=False)
+    assert result, (
+        "Ref YML and parsed YML\
+has not the same structure!!"
+    )
+    os.remove(test_yml_file)
[…]
