+++ tests/test_nyaml2nxdl.py
@@ -16,68 +16,32 @@
 # limitations under the License.
 #
 """
-Tests for nyaml2nxdl tool
+Tests for the YAML → NXDL conversion direction (nyaml2nxdl).
 """
 
-import filecmp
 import os
-import re
 import sys
 from collections import OrderedDict
-from datetime import datetime
 from pathlib import Path
 
 import lxml.etree as ET
 import pytest
 import yaml
 from click.testing import CliRunner
+from helpers import check_and_replace_latest_copyright, check_file_fresh_baked
 
 from nyaml import cli as nyaml2nxdl
-from nyaml import nyaml2nxdl as nyaml2nxdl_forward_tools
-from nyaml.comment_collector import CommentCollector
-from nyaml.helper import LineLoader, remove_namespace_from_tag
-from nyaml.nyaml2nxdl import get_nxdl_copyright_license, handle_each_part_doc
-
-LATEST_COPYRIGHT_YEAR = f"{datetime.now().year}-{datetime.now().year}"
-LATEST_COPYRIGHT = rf"# Copyright \(C\) {LATEST_COPYRIGHT_YEAR} NeXus International Advisory Committee \(NIAC\)"
-COPYRIGHT_REPLACEMENT = (
-    f"# Copyright (C) 2010-2020 NeXus International Advisory Committee (NIAC)"
-)
-
-
-def check_and_replace_latest_copyright(nxdl_file):
-    """Check if the latest copyright date has been written properly,
-    and replace it accordingly to the test data file.
-    """
-    content = nxdl_file.read_text()
-    generated_copyright = re.findall(LATEST_COPYRIGHT, content, re.DOTALL)
-    assert len(generated_copyright) == 1, (
-        f"No copyright or not correct copyright year found in {nxdl_file}"
-    )
-    content = re.sub(LATEST_COPYRIGHT, COPYRIGHT_REPLACEMENT, content)
-    nxdl_file.write_text(content)
+from nyaml.helper import remove_namespace_from_tag
+from nyaml.nyaml2nxdl import handle_each_part_doc
 
 
 def delete_duplicates(list_of_matching_string):
-    """
-    Delete duplicate from lists
-    """
+    """Delete duplicates from a list while preserving order."""
     return list(dict.fromkeys(list_of_matching_string))
 
 
-def check_file_fresh_baked(test_file):
-    """
-    Get sure that the test file is generated by the converter
-    """
-    path = Path(test_file)
-    timestamp = datetime.fromtimestamp(path.stat().st_mtime).strftime("%d/%m/%Y %H:%M")
-    now = datetime.now().strftime("%d/%m/%Y %H:%M")
-    assert timestamp == now, "xml file not generated"
-
-
 def find_matches(xml_file, desired_matches):
-    """
-        Read xml file and find desired matches. Return a list of two lists in the form:
+    """Read xml file and find desired matches. Return a list of two lists in the form:
     [[matching_line],[matching_line_index]]
     """
     with open(xml_file, encoding="utf-8") as file:
@@ -100,9 +64,8 @@ def find_matches(xml_file, desired_matches):
 
 
 def compare_matches(ref_xml_file, test_yml_file, test_xml_file, desired_matches):
-    """
-        Check if a new xml file is generated
-    and if test xml file is equal to reference xml file
+    """Check if a new xml file is generated
+    and if test xml file is equal to reference xml file.
     """
     # Reference file is read
[…]

[…]
