+++ tests/data/nxdl2yaml/ref_doc_nxdl2yaml.yaml
@@ -1,5 +1,5 @@
-category: application
-doc: |
+\category: application
+\doc: |
   This is a definition for data to be archived by ICAT (http://www.icatproject.org/).
   
   .. text from the icatproject.org site
@@ -8,9 +8,9 @@ doc: |
           interface to all ISIS experimental data and will provide
           a mechanism to link all aspects of ISIS research from
           proposal through to publication.
-type: group
+\type: group
 NXarchive(NXobject):
-  (NXentry)entry:
+  entry(NXentry):
     \@index:
     title:
     experiment_identifier(NX_CHAR):
@@ -64,100 +64,3 @@ NXarchive(NXobject):
       \doc: |
         unique identifier for this measurement as provided by the facility
     start_time(NX_DATE_TIME):
-
-# ++++++++++++++++++++++++++++++++++ SHA HASH ++++++++++++++++++++++++++++++++++
-# 558e3478f110c46ffd79754e086388960603584a7f8c08ee6f697f9dbbc78c69
-# <?xml version='1.0' encoding='UTF-8'?>
-# <?xml-stylesheet type="text/xsl" href="nxdlformat.xsl"?>
-# <!--
-# # NeXus - Neutron and X-ray Common Data Format
-# #
-# # Copyright (C) 2010-2020 NeXus International Advisory Committee (NIAC)
-# #
-# # This library is free software; you can redistribute it and/or
-# # modify it under the terms of the GNU Lesser General Public
-# # License as published by the Free Software Foundation; either
-# # version 3 of the License, or (at your option) any later version.
-# #
-# # This library is distributed in the hope that it will be useful,
-# # but WITHOUT ANY WARRANTY; without even the implied warranty of
-# # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# # Lesser General Public License for more details.
-# #
-# # You should have received a copy of the GNU Lesser General Public
-# # License along with this library; if not, write to the Free Software
-# # Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
-# #
-# # For further information, see http://www.nexusformat.org
-# -->
-# <definition xmlns="http://definition.nexusformat.org/nxdl/3.1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" category="application" type="group" name="NXarchive" extends="NXobject" xsi:schemaLocation="http://definition.nexusformat.org/nxdl/3.1 ../nxdl.xsd">
-#   <doc>
-#        This is a definition for data to be archived by ICAT (http://www.icatproject.org/).
-# 
-#        .. text from the icatproject.org site
-# 
-#        		the database (with supporting software) that provides an
-#        		interface to all ISIS experimental data and will provide
-#        		a mechanism to link all aspects of ISIS research from
-#        		proposal through to publication.
-#   </doc>
-#   <group type="NXentry" name="entry">
-#     <attribute name="index"/>
-#     <field name="title"/>
-#     <field name="experiment_identifier" type="NX_CHAR">
-#       <doc>
-#            unique identifier for the experimenta
-#              lkokö
-#              sddfdfd
-# 
-#              This concept is related to term `12.58`_ of the ISO 18115-1:2023 standard.
-# 
-#              .. _12.58: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:term:12.58
-# 
-#            This is test doc.
-#       </doc>
-#     </field>
-#     <field name="experiment_description" type="NX_CHAR">
-#       <doc>
-#            Part1: Text line 1.
-#            part1: Text line 2
-# 
[…]

[…]
