### 2026-04-29 e3d0835
Split tests in forward, backward and roundtrip direction.
