+++ tests/test_nxdl2nyaml.py
@@ -77,9 +77,9 @@ def test_nxdl2yaml_doc_format_and_nxdl_part_as_comment():
     to yaml to check if they are correct.
             2. In test-2: Check the nxdl that comes at the end of yaml file as comment.
     """
-    ref_xml_file = str(NXDL2NYAML_DATA_DIR / "Ref_NXentry.nxdl.xml")
-    ref_yml_file = str(NXDL2NYAML_DATA_DIR / "Ref_NXentry.yaml")
-    test_yml_file = str(NXDL2NYAML_DATA_DIR / "Ref_NXentry_parsed.yaml")
+    ref_xml_file = "tests/data/nxdl2yaml/Ref_NXentry.nxdl.xml"
+    ref_yml_file = "tests/data/nxdl2yaml/Ref_NXentry.yaml"
+    test_yml_file = "tests/data/nxdl2yaml/Ref_NXentry_parsed.yaml"
     result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
     assert result.exit_code == 0
     check_file_fresh_baked(test_yml_file)
@@ -118,9 +118,10 @@ def test_nxdl2yaml_enumerations():
 def test_nxdl2yaml_doc():
     """To test the doc style from nxdl to yaml."""
 
-    nxdl_file = str(NXDL2NYAML_DATA_DIR / "doc_nxdl2yaml.nxdl.xml")
-    ref_yaml = str(NXDL2NYAML_DATA_DIR / "ref_doc_nxdl2yaml.yaml")
-    parsed_yaml_file = str(NXDL2NYAML_DATA_DIR / "doc_nxdl2yaml_parsed.yaml")
+    pwd = Path(__file__).parent
+    nxdl_file = pwd / "data/nxdl2yaml/doc_nxdl2yaml.nxdl.xml"
+    ref_yaml = pwd / "data/nxdl2yaml/ref_doc_nxdl2yaml.yaml"
+    parsed_yaml_file = pwd / "data/nxdl2yaml/doc_nxdl2yaml_parsed.yaml"
 
     result = CliRunner().invoke(
         nyaml2nxdl.launch_tool, ["--do-not-store-nxdl", str(nxdl_file)]
