### 2026-04-29 ed03690
Split tests in forward, backward and roundtrip direction.
