+++ tests/test_nyaml2nxdl.py
@@ -34,6 +34,8 @@ from nyaml import cli as nyaml2nxdl
 from nyaml.helper import remove_namespace_from_tag
 from nyaml.nyaml2nxdl import handle_each_part_doc
 
+FORWARD_DATA_DIR = "tests/data/yaml2nxdl"
+
 
 def delete_duplicates(list_of_matching_string):
     """Delete duplicates from a list while preserving order."""
@@ -83,9 +85,9 @@ def test_links():
     """
     Check the correct parsing of links
     """
-    ref_xml_link_file = "tests/data/Ref_NXtest_links.nxdl.xml"
-    test_yml_link_file = "tests/data/NXtest_links.yaml"
-    test_xml_link_file = "tests/data/NXtest_links.nxdl.xml"
+    ref_xml_link_file = f"{FORWARD_DATA_DIR}/Ref_NXtest_links.nxdl.xml"
+    test_yml_link_file = f"{FORWARD_DATA_DIR}/NXtest_links.yaml"
+    test_xml_link_file = f"{FORWARD_DATA_DIR}/NXtest_links.nxdl.xml"
     # ref_xml_link_file = os.path.abspath(data_path + '/Ref_NXtest_links.nxdl.xml')
     # test_yml_link_file = os.path.abspath(data_path + '/NXtest_links.yaml')
     # test_xml_link_file = os.path.abspath(data_path + '/NXtest_links.nxdl.xml')
@@ -93,7 +95,7 @@ def test_links():
     compare_matches(
         ref_xml_link_file, test_yml_link_file, test_xml_link_file, desired_matches
     )
-    os.remove("tests/data/NXtest_links.nxdl.xml")
+    os.remove(f"{FORWARD_DATA_DIR}/NXtest_links.nxdl.xml")
     sys.stdout.write("Test on links okay.\n")
 
 
@@ -293,27 +295,27 @@ def test_file_line_error():
     """
     In this test the yaml file line in the error message is tested.
     """
-    test_yml_file = "tests/data/NXfilelineError1.yaml"
-    out_nxdl = "tests/data/NXfilelineError1.nxdl.xml"
-    out_yaml = "tests/data/temp_NXfilelineError1.yaml"
+    test_yml_file = f"{FORWARD_DATA_DIR}/NXfilelineError1.yaml"
+    out_nxdl = f"{FORWARD_DATA_DIR}/NXfilelineError1.nxdl.xml"
+    out_yaml = f"{FORWARD_DATA_DIR}/temp_NXfilelineError1.yaml"
     result = CliRunner().invoke(nyaml2nxdl.launch_tool, [test_yml_file])
     assert result.exit_code == 1
     assert "13" in str(result.exception)
     os.remove(out_nxdl)
     os.remove(out_yaml)
 
-    test_yml_file = "tests/data/NXfilelineError2.yaml"
-    out_nxdl = "tests/data/NXfilelineError2.nxdl.xml"
-    out_yaml = "tests/data/temp_NXfilelineError2.yaml"
+    test_yml_file = f"{FORWARD_DATA_DIR}/NXfilelineError2.yaml"
+    out_nxdl = f"{FORWARD_DATA_DIR}/NXfilelineError2.nxdl.xml"
+    out_yaml = f"{FORWARD_DATA_DIR}/temp_NXfilelineError2.yaml"
     result = CliRunner().invoke(nyaml2nxdl.launch_tool, [test_yml_file])
     assert result.exit_code == 1
     assert "21" in str(result.exception)
     os.remove(out_nxdl)
[…]
