### 2026-04-29 a74b01c
further fixes for backslash in root
