+++ tests/test_yaml2nxdl.py
@@ -1,578 +1,0 @@
-#
-# Copyright The NOMAD Authors.
-#
-# This file is part of NOMAD. See https://nomad-lab.eu for further info.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-"""
-Tests for nyaml -> nxdl conversion (yaml2nxdl direction).
-"""
-
-import os
-import re
-import sys
-from collections import OrderedDict
-from datetime import datetime
-from pathlib import Path
-
-import lxml.etree as ET
-import pytest
-import yaml
-from click.testing import CliRunner
-
-from nyaml import cli as nyaml_cli
-from nyaml.helper import remove_namespace_from_tag
-from nyaml.nyaml2nxdl import get_nxdl_copyright_license, handle_each_part_doc
-
-LATEST_COPYRIGHT_YEAR = f"{datetime.now().year}-{datetime.now().year}"
-LATEST_COPYRIGHT = rf"# Copyright \(C\) {LATEST_COPYRIGHT_YEAR} NeXus International Advisory Committee \(NIAC\)"
-COPYRIGHT_REPLACEMENT = (
-    f"# Copyright (C) 2010-2020 NeXus International Advisory Committee (NIAC)"
-)
-
-DATA = Path(__file__).parent / "data" / "yaml2nxdl"
-
-
-def check_and_replace_latest_copyright(nxdl_file):
-    content = nxdl_file.read_text()
-    generated_copyright = re.findall(LATEST_COPYRIGHT, content, re.DOTALL)
-    assert len(generated_copyright) == 1, (
-        f"No copyright or not correct copyright year found in {nxdl_file}"
-    )
-    content = re.sub(LATEST_COPYRIGHT, COPYRIGHT_REPLACEMENT, content)
-    nxdl_file.write_text(content)
-
-
-def delete_duplicates(list_of_matching_string):
-    return list(dict.fromkeys(list_of_matching_string))
-
-
-def check_file_fresh_baked(test_file):
-    path = Path(test_file)
-    timestamp = datetime.fromtimestamp(path.stat().st_mtime).strftime("%d/%m/%Y %H:%M")
-    now = datetime.now().strftime("%d/%m/%Y %H:%M")
-    assert timestamp == now, "xml file not generated"
-
-
-def find_matches(xml_file, desired_matches):
-    with open(xml_file, encoding="utf-8") as file:
-        xml_reference = file.readlines()
-    lines = []
-    lines_index = []
-    found_matches = []
-    for i, line in enumerate(xml_reference):
-        for desired_match in desired_matches:
-            if str(desired_match) in str(line):
-                lines.append(line)
-                lines_index.append(i)
-                found_matches.append(desired_match)
-    found_matches_clean = delete_duplicates(found_matches)
-    assert len(found_matches_clean) == len(desired_matches), (
-        "some desired_matches were \nnot found in file"
-    )
-    return [lines, lines_index]
-
-
-def compare_matches(ref_xml_file, test_yml_file, test_xml_file, desired_matches):
-    ref_matches = find_matches(ref_xml_file, desired_matches)
-    runner = CliRunner()
-    result = runner.invoke(nyaml_cli.launch_tool, [test_yml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_xml_file)
[…]

[…]
