+++ tests/data/nxdl2yaml/ref_doc_nxdl2yaml.yaml
@@ -1,5 +1,5 @@
-\category: application
-\doc: |
+category: application
+doc: |
   This is a definition for data to be archived by ICAT (http://www.icatproject.org/).
   
   .. text from the icatproject.org site
@@ -8,13 +8,13 @@
           interface to all ISIS experimental data and will provide
           a mechanism to link all aspects of ISIS research from
           proposal through to publication.
-\type: group
+type: group
 NXarchive(NXobject):
-  entry(NXentry):
+  (NXentry)entry:
     \@index:
     title:
     experiment_identifier(NX_CHAR):
-      \doc:
+      doc:
       - |
         unique identifier for the experimenta
           lkokö
@@ -27,7 +27,7 @@ NXarchive(NXobject):
       - |
         This is test doc.
     experiment_description(NX_CHAR):
-      \doc:
+      doc:
       - |
         Part1: Text line 1.
         part1: Text line 2
@@ -37,13 +37,13 @@ NXarchive(NXobject):
           term: term
           url: url
     only_iso_string(NX_CHAR):
-      \doc: |
+      doc: |
         xref:
           spec: ISO 18115-1:2023
           term: 1.1
           url: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:term:1.1
     double_iso_string(NX_CHAR):
-      \doc:
+      doc:
       - |
         Way of scanning the energy axis (fixed or sweep).
       - |
@@ -57,10 +57,107 @@ NXarchive(NXobject):
           term: 12.66
           url: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:term:12.66
     collection_identifier(NX_CHAR):
-      \doc: |
+      doc: |
         ID of user or DAQ define group of data files
         Brief summary of the collection, including grouping criteria
     entry_identifier(NX_CHAR):
-      \doc: |
+      doc: |
         unique identifier for this measurement as provided by the facility
     start_time(NX_DATE_TIME):
+
+# ++++++++++++++++++++++++++++++++++ SHA HASH ++++++++++++++++++++++++++++++++++
+# 558e3478f110c46ffd79754e086388960603584a7f8c08ee6f697f9dbbc78c69
+# <?xml version='1.0' encoding='UTF-8'?>
+# <?xml-stylesheet type="text/xsl" href="nxdlformat.xsl"?>
+# <!--
+# # NeXus - Neutron and X-ray Common Data Format
+# #
+# # Copyright (C) 2010-2020 NeXus International Advisory Committee (NIAC)
+# #
+# # This library is free software; you can redistribute it and/or
+# # modify it under the terms of the GNU Lesser General Public
+# # License as published by the Free Software Foundation; either
+# # version 3 of the License, or (at your option) any later version.
+# #
+# # This library is distributed in the hope that it will be useful,
+# # but WITHOUT ANY WARRANTY; without even the implied warranty of
+# # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# # Lesser General Public License for more details.
+# #
+# # You should have received a copy of the GNU Lesser General Public
+# # License along with this library; if not, write to the Free Software
+# # Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+# #
+# # For further information, see http://www.nexusformat.org
+# -->
+# <definition xmlns="http://definition.nexusformat.org/nxdl/3.1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" category="application" type="group" name="NXarchive" extends="NXobject" xsi:schemaLocation="http://definition.nexusformat.org/nxdl/3.1 ../nxdl.xsd">
+#   <doc>
[…]
