+++ tests/test_nxdl2nyaml.py
@@ -26,10 +26,16 @@ from pathlib import Path
 
 import pytest
 from click.testing import CliRunner
-from helpers import check_file_fresh_baked, compare_yaml_content
+from helpers import (
+    check_file_fresh_baked,
+    compare_log_and_reference,
+    compare_yaml_content,
+)
 
 from nyaml import cli as nyaml2nxdl
 from nyaml.helper import LineLoader
+
+NXDL2NYAML_TESTS_DIR = Path(__file__).parent / "data" / "nxdl2yaml"
 
 
 def test_nxdl2yaml_nameType():
@@ -37,9 +43,9 @@ def test_nxdl2yaml_nameType():
     Check the correct handling of the nameType attribute for the direction
     nxdl->nyaml.
     """
-    ref_xml_file = "tests/data/nxdl2yaml/allowed_nameType.nxdl.xml"
-    ref_yml_file = "tests/data/nxdl2yaml/ref_allowed_nameType.yaml"
-    test_yml_file = "tests/data/nxdl2yaml/allowed_nameType_parsed.yaml"
+    ref_xml_file = str(NXDL2NYAML_TESTS_DIR / "allowed_nameType.nxdl.xml")
+    ref_yml_file = str(NXDL2NYAML_TESTS_DIR / "ref_allowed_nameType.yaml")
+    test_yml_file = str(NXDL2NYAML_TESTS_DIR / "allowed_nameType_parsed.yaml")
 
     result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
     assert result.exit_code == 0
@@ -64,6 +70,29 @@ has not the same structure!!"
     sys.stdout.write("Test on xml -> yml doc formatting okay.\n")
 
 
+@pytest.mark.parametrize(
+    "test_input",
+    [
+        ("NXdimensionsType"),
+    ],
+)
+def test_dimension(test_input):
+    """
+    Tests if the conversion of specific test files from NXDL to YAML results as expected.
+    Expected output files shall have the corresponding name with the prefix Ref_.
+    """
+    test_xml_input_file = str(NXDL2NYAML_TESTS_DIR / f"{test_input}.nxdl.xml")
+    test_yml_output_file = str(NXDL2NYAML_TESTS_DIR / f"{test_input}_parsed.yaml")
+    ref_yml_output_file = str(NXDL2NYAML_TESTS_DIR / f"Ref_{test_input}.yaml")
+    runner = CliRunner()
+    result = runner.invoke(nyaml2nxdl.launch_tool, [test_xml_input_file])
+    assert result.exit_code == 0
+
+    compare_log_and_reference(test_yml_output_file, ref_yml_output_file)
+
+    os.remove(test_yml_output_file)
+
+
 def test_nxdl2yaml_doc_format_and_nxdl_part_as_comment():
     """
     This test for two reasons:
@@ -71,9 +100,9 @@ def test_nxdl2yaml_doc_format_and_nxdl_part_as_comment():
     to yaml to check if they are correct.
             2. In test-2: Check the nxdl that comes at the end of yaml file as comment.
     """
-    ref_xml_file = "tests/data/nxdl2yaml/Ref_NXentry.nxdl.xml"
-    ref_yml_file = "tests/data/nxdl2yaml/Ref_NXentry.yaml"
-    test_yml_file = "tests/data/nxdl2yaml/Ref_NXentry_parsed.yaml"
+    ref_xml_file = str(NXDL2NYAML_TESTS_DIR / "Ref_NXentry.nxdl.xml")
[…]
