### 2026-05-04 e16401a
Reuse/modify/merge unused test files
what happened to each file:

File — Action — Reason
tests/data/nxdl2yaml/dimensions.nxdl.xml — Merged + deleted — Had one unique case (ref attribute on <dim>); that case was folded into NXdimensionsType.nxdl.xml as a new dim_with_ref field. The rest was duplicate coverage.
tests/data/nxdl2yaml/ref_dimensions.yaml — Deleted — Reference file for dimensions.nxdl.xml; no longer needed after the merge.
tests/data/unused_files/NXattributes.nxdl.xml — Moved to nxdl2yaml/ — Was never tested despite being a valid backward-direction test file. Moved and wired up in t
[…]
