### 2026-05-04 6d3431b
update documentation with reserved keywords
