+++ docs/tutorials/writing-a-yaml-definition.md
@@ -13,11 +13,11 @@ You will know
 
 - the basic structure of a NeXus definition written in YAML
 - how to write groups, fields, and attributes
-- how to use the key special keywords (`exists`, `unit`, `dimensions`, `enumeration`)
+- how to use the key special keywords (`\exists`, `\unit`, `\dimensions`, `\enumeration`, etc.)
 - how to convert a YAML definition to NXDL XML using `nyaml2nxdl`
 
 !!! note
-    We are making extensive use of special keywords here (`nameType`, `exists`, `unit`, `dimensions`, `enumeration`, `xref`) that indicate specific parts of the NeXus schema in the YAML notation. To learn more, see [Learn > nyaml notation](../learn/yaml-notation.md). The NeXus `choice` concept is not yet supported in `nyaml`.
+    We are making extensive use of special keywords here (`\nameType`, `\exists`, `\unit`, `\dimensions`, `\enumeration`, `\xref`) that indicate specific parts of the NeXus schema in the YAML notation. All of these keywords are identified by the `\` prefix. To learn more, see [Learn > nyaml notation](../learn/yaml-notation.md). The NeXus `choice` concept is not yet supported in `nyaml`.
 
 ---
 
@@ -29,15 +29,15 @@ We will build `NXmy_measurement`, a minimal but complete application definition 
 
 ## Step 1: The root section
 
-Every NeXus definition starts with a root section. It sets the `category` (`application` or `base`), the `type` (always `group`), a human-readable `doc` string, and the name of the NeXus definition.
+Every NeXus definition starts with a root section. It sets the `\category` (`application` or `base`), the `\type` (always `group`), a human-readable `\doc` string, and the name of the NeXus definition.
 
 Create a file called `NXmy_measurement.yaml` with this content:
 
 ```yaml
-category: application
-doc: |
+\category: application
+\doc: |
   Minimal application definition for a 2D detector measurement.
-type: group
+\type: group
 NXmy_measurement(NXobject):
 ```
 
@@ -61,21 +61,21 @@ Real NeXus application definitions wrap their data in an `NXentry` group. Inside
 NXmy_measurement(NXobject):
   (NXentry):
     definition:
-      enumeration: [NXmy_measurement]
+      \enumeration: [NXmy_measurement]
     title:
     start_time(NX_DATE_TIME):
-      doc: Datetime of the start of the measurement.
+      \doc: Datetime of the start of the measurement.
     end_time(NX_DATE_TIME):
-      exists: recommended
-      doc: Datetime of the end of the measurement.
+      \exists: recommended
+      \doc: Datetime of the end of the measurement.
 ```
 
 A few things to notice:
 
-- `(NXentry):` without a name means any valid group name is accepted in the HDF5 file (e.g. `entry`, `entry1`). This is `nameType: any`, the default for anonymous groups.
+- `(NXentry):` without a name means any valid group name is accepted in the HDF5 file (e.g. `entry`, `entry1`). This is `\nameType: any`, the default for anonymous groups.
 - `definition:` is a field with no type, so it defaults to `NX_CHAR`.
 - `start_time(NX_DATE_TIME):` declares a field with NeXus type `NX_DATE_TIME`.
-- `exists: recommended` on `end_time` means it is advised but not required. By default all concepts in an application definition are required; `exists` relaxes that. In base classes, however, all concepts are optional.
+- `\exists: recommended` on `end_time` means it is advised but not required. By default all concepts in an application definition are required; `\exists` relaxes that. In base classes, however, all concepts are optional.
 
 ---
 
@@ -87,63 +87,63 @@ The `NXinstrument` group describes the experimental setup. Inside it we add a de
 NXmy_measurement(NXobject):
   (NXentry):
[…]
