### 2026-05-04 b204b74
\target, \xref, \open_enum in tests
