+++ tests/data/nxdl2yaml/ref_allowed_nameType.yaml
@@ -13,7 +13,7 @@ NXallowed_name_type(NXobject):
         \doc: |
           An all lower case attribute with no nameType.
     lower_case_link_no_nametype(link):
-      target: field
+      \target: field
       \doc: |
         An all lower case link with no nameType.
   lower_case_group_specified(NXobject):
@@ -29,7 +29,7 @@ NXallowed_name_type(NXobject):
         \doc: |
           An all lower case attribute with nameType="specified".
       lower_case_link_specified(link):
-        target: field
+        \target: field
         \doc: |
           An all lower case link with no nameType.
     (NXobject):
@@ -42,7 +42,7 @@ NXallowed_name_type(NXobject):
           \doc: |
             An unnamed attribute no nameType. Interpretead as nameType="any".
         LINK(link):
-          target: field
+          \target: field
           \doc: |
             An unnamed link with no nameType. Interpretead as nameType="any".
     (NXdata):
@@ -59,7 +59,7 @@ NXallowed_name_type(NXobject):
             An unnamed attribute with nameType="any".
         LINK(link):
           \nameType: any
-          target: field
+          \target: field
           \doc: |
             An unnamed link with nameType="any".
     objectOBJECT(NXobject):
@@ -76,7 +76,7 @@ NXallowed_name_type(NXobject):
             A partially-renameable attribute nameType="partial". Variadic part at the end.
         linkLINK(link):
           \nameType: partial
-          target: field
+          \target: field
           \doc: |
             A partially-renameable link nameType="partial". Variadic part at the end.
     objectOBJECTobject(NXobject):
@@ -95,7 +95,7 @@ NXallowed_name_type(NXobject):
             A partially-renameable attribute with nameType="partial". Variadic part in the
             middle.
         linkLINKlink(link):
-          target: field
+          \target: field
           \nameType: partial
           \doc: |
             A partially-renameable link with nameType="partial". Variadic part in the
@@ -116,7 +116,7 @@ NXallowed_name_type(NXobject):
             A partially-renameable attribute with nameType="partial". Variadic part at the
             start.
         LINKlink(link):
-          target: field
+          \target: field
           \nameType: partial
           \doc: |
             A partially-renameable link with nameType="partial". Variadic part at the start.
@@ -136,7 +136,7 @@ NXallowed_name_type(NXobject):
             A partially-renameable attribute with nameType="partial". Multiple variadic and
             non-variadic parts.
         LINKlinkLINKlink(link):
-          target: field
+          \target: field
           \nameType: specified
           \doc: |
             A partially-renameable link with nameType="partial". Multiple variadic and non-
