+++ tests/data/yaml2nxdl/NXtest_links.yaml
@@ -3,7 +3,7 @@
 NXentry:
    (NXdata):
      polar_angle(link):
-       target: here1
+       \target: here1
        \nameType: specified
      target_angle(link):
-       target: here2
+       \target: here2

[…]
