### 2026-05-04 3476ea4
implement \target, \xref, \open (for enums) as reserved keywords
