+++ tests/test_nyaml2nxdl.py
@@ -28,66 +28,22 @@ import lxml.etree as ET
 import pytest
 import yaml
 from click.testing import CliRunner
-from helpers import check_and_replace_latest_copyright, check_file_fresh_baked
+from helpers import check_and_replace_latest_copyright, compare_matches
 
 from nyaml import cli as nyaml2nxdl
 from nyaml.helper import remove_namespace_from_tag
 from nyaml.nyaml2nxdl import handle_each_part_doc
 
-FORWARD_DATA_DIR = Path(__file__).parent / "data" / "yaml2nxdl"
-
-
-def delete_duplicates(list_of_matching_string):
-    """Delete duplicates from a list while preserving order."""
-    return list(dict.fromkeys(list_of_matching_string))
-
-
-def find_matches(xml_file, desired_matches):
-    """Read xml file and find desired matches. Return a list of two lists in the form:
-    [[matching_line],[matching_line_index]]
-    """
-    with open(xml_file, encoding="utf-8") as file:
-        xml_reference = file.readlines()
-    lines = []
-    lines_index = []
-    found_matches = []
-    for i, line in enumerate(xml_reference):
-        for desired_match in desired_matches:
-            if str(desired_match) in str(line):
-                lines.append(line)
-                lines_index.append(i)
-                found_matches.append(desired_match)
-    # ascertain that all the desired matches were found in file
-    found_matches_clean = delete_duplicates(found_matches)
-    assert len(found_matches_clean) == len(desired_matches), (
-        "some desired_matches were \nnot found in file"
-    )
-    return [lines, lines_index]
-
-
-def compare_matches(ref_xml_file, test_yml_file, test_xml_file, desired_matches):
-    """Check if a new xml file is generated
-    and if test xml file is equal to reference xml file.
-    """
-    # Reference file is read
-    ref_matches = find_matches(ref_xml_file, desired_matches)
-    # Test file is generated
-    runner = CliRunner()
-    result = runner.invoke(nyaml2nxdl.launch_tool, [test_yml_file])
-    assert result.exit_code == 0
-    check_file_fresh_baked(test_xml_file)
-    # Test file is read
-    test_matches = find_matches(test_xml_file, desired_matches)
-    assert test_matches == ref_matches
+NYAML2NXDL_DATA_DIR = Path(__file__).parent / "data" / "yaml2nxdl"
 
 
 def test_links():
     """
     Check the correct parsing of links
     """
-    ref_xml_link_file = str(FORWARD_DATA_DIR / "Ref_NXtest_links.nxdl.xml")
-    test_yml_link_file = str(FORWARD_DATA_DIR / "NXtest_links.yaml")
-    test_xml_link_file = str(FORWARD_DATA_DIR / "NXtest_links.nxdl.xml")
+    ref_xml_link_file = str(NYAML2NXDL_DATA_DIR / "Ref_NXtest_links.nxdl.xml")
+    test_yml_link_file = str(NYAML2NXDL_DATA_DIR / "NXtest_links.yaml")
+    test_xml_link_file = str(NYAML2NXDL_DATA_DIR / "NXtest_links.nxdl.xml")
[…]
