+++ tests/test_nxdl2nyaml.py
@@ -35,7 +35,7 @@ from helpers import (
 from nyaml import cli as nyaml2nxdl
 from nyaml.helper import LineLoader
 
-NXDL2NYAML_TESTS_DIR = Path(__file__).parent / "data" / "nxdl2yaml"
+NXDL2NYAML_DATA_DIR = Path(__file__).parent / "data" / "nxdl2yaml"
 
 
 def test_nxdl2yaml_nameType():
@@ -43,9 +43,9 @@ def test_nxdl2yaml_nameType():
     Check the correct handling of the nameType attribute for the direction
     nxdl->nyaml.
     """
-    ref_xml_file = str(NXDL2NYAML_TESTS_DIR / "allowed_nameType.nxdl.xml")
-    ref_yml_file = str(NXDL2NYAML_TESTS_DIR / "ref_allowed_nameType.yaml")
-    test_yml_file = str(NXDL2NYAML_TESTS_DIR / "allowed_nameType_parsed.yaml")
+    ref_xml_file = str(NXDL2NYAML_DATA_DIR / "allowed_nameType.nxdl.xml")
+    ref_yml_file = str(NXDL2NYAML_DATA_DIR / "ref_allowed_nameType.yaml")
+    test_yml_file = str(NXDL2NYAML_DATA_DIR / "allowed_nameType_parsed.yaml")
 
     result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
     assert result.exit_code == 0
@@ -81,9 +81,9 @@ def test_dimension(test_input):
     Tests if the conversion of specific test files from NXDL to YAML results as expected.
     Expected output files shall have the corresponding name with the prefix Ref_.
     """
-    test_xml_input_file = str(NXDL2NYAML_TESTS_DIR / f"{test_input}.nxdl.xml")
-    test_yml_output_file = str(NXDL2NYAML_TESTS_DIR / f"{test_input}_parsed.yaml")
-    ref_yml_output_file = str(NXDL2NYAML_TESTS_DIR / f"Ref_{test_input}.yaml")
+    test_xml_input_file = str(NXDL2NYAML_DATA_DIR / f"{test_input}.nxdl.xml")
+    test_yml_output_file = str(NXDL2NYAML_DATA_DIR / f"{test_input}_parsed.yaml")
+    ref_yml_output_file = str(NXDL2NYAML_DATA_DIR / f"Ref_{test_input}.yaml")
     runner = CliRunner()
     result = runner.invoke(nyaml2nxdl.launch_tool, [test_xml_input_file])
     assert result.exit_code == 0
@@ -100,9 +100,9 @@ def test_nxdl2yaml_doc_format_and_nxdl_part_as_comment():
     to yaml to check if they are correct.
             2. In test-2: Check the nxdl that comes at the end of yaml file as comment.
     """
-    ref_xml_file = str(NXDL2NYAML_TESTS_DIR / "Ref_NXentry.nxdl.xml")
-    ref_yml_file = str(NXDL2NYAML_TESTS_DIR / "Ref_NXentry.yaml")
-    test_yml_file = str(NXDL2NYAML_TESTS_DIR / "Ref_NXentry_parsed.yaml")
+    ref_xml_file = str(NXDL2NYAML_DATA_DIR / "Ref_NXentry.nxdl.xml")
+    ref_yml_file = str(NXDL2NYAML_DATA_DIR / "Ref_NXentry.yaml")
+    test_yml_file = str(NXDL2NYAML_DATA_DIR / "Ref_NXentry_parsed.yaml")
     result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
[…]
