### 2026-05-05 cc05dda
Split tests in forward, backward and roundtrip direction.  (#92)
* Reactor the code.

* move round-trip tiles.

* Split tests in forward, backward, and round-trip directions.

* Stored unused files in a folder.

* Update the test file path.

* fix file path.

* Update the test data path and keep the non-test function in the helper file.

* fix test.
