+++ tests/test_roundtrip.py
@@ -1,0 +1,219 @@
+#
+# Copyright The NOMAD Authors.
+#
+# This file is part of NOMAD. See https://nomad-lab.eu for further info.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""
+Tests for bidirectional (roundtrip) conversions: nxdl→yaml→nxdl and yaml→nxdl→yaml.
+"""
+
+import os
+import re
+import sys
+from pathlib import Path
+
+import lxml.etree as ET
+from click.testing import CliRunner
+from helpers import check_file_fresh_baked, compare_yaml_content
+
+from nyaml import cli as nyaml2nxdl
+from nyaml import nyaml2nxdl as nyaml2nxdl_forward_tools
+from nyaml.comment_collector import CommentCollector
+from nyaml.helper import LineLoader
+from nyaml.nyaml2nxdl import get_nxdl_copyright_license
+
+ROUNDTRIP_DATA_DIR = Path(__file__).parent / "data" / "roundtrip"
+
+
+def test_xml_parsing():
+    """In this test an xml file is converted to yml and then back to xml.
+    The xml trees of the two files are then compared.
+    """
+    ref_xml_file = str(ROUNDTRIP_DATA_DIR / "Ref_NXellipsometry.nxdl.xml")
+    test_yml_file = str(ROUNDTRIP_DATA_DIR / "Ref_NXellipsometry_parsed.yaml")
+    test_xml_file = str(ROUNDTRIP_DATA_DIR / "Ref_NXellipsometry_parsed.nxdl.xml")
+    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_xml_file])
+    assert result.exit_code == 0
+    check_file_fresh_baked(test_yml_file)
+    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [test_yml_file])
+    assert result.exit_code == 0
+    check_file_fresh_baked(test_xml_file)
+
+    test_tree = ET.parse(test_xml_file)
+    test_tree_flattened = {i.tag.split("}", 1)[1] for i in test_tree.iter()}
+
+    ref_tree = ET.parse(ref_xml_file)
+    ref_tree_flattened = {i.tag.split("}", 1)[1] for i in ref_tree.iter()}
+
+    assert test_tree_flattened == ref_tree_flattened, (
+        "Ref XML and parsed XML\
+has not the same tree structure!!"
+    )
+    os.remove(test_xml_file)
+    os.remove(test_yml_file)
+    sys.stdout.write("Test on xml -> yml -> xml okay.\n")
+
+
+def test_yml_parsing():
+    """In this test a yml file is converted to xml and then back to yml.
+    The yml trees of the two files are then compared.
+    """
+    ref_yml_file = str(ROUNDTRIP_DATA_DIR / "Ref_NXellipsometry.yaml")
+    test_xml_file = str(ROUNDTRIP_DATA_DIR / "Ref_NXellipsometry.nxdl.xml")
+    test_yml_file = str(ROUNDTRIP_DATA_DIR / "Ref_NXellipsometry_parsed.yaml")
+    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [ref_yml_file])
+    assert result.exit_code == 0
[…]
