### 2026-05-05 d688705
remove unused dim_parameters
