+++ tests/data/yaml2nxdl/NXattributes.yaml
@@ -37,8 +37,6 @@ NXellipsometry_base_draft(my_test_extends):
       \dimensions:
         \rank: 3
         \dim: [[3, N_calibration_angles+1], [2, N_variables], [1, N_calibration_wavelength]]
-        dim_parameters:
-          required: ['true', 'true', 'true']
     test_field:
       \@vector(NX_NUMBER):
         \doc: Attribute with specific datatype within a field.
