### 2026-05-05 67fcb0e
add a docs section on NXDL coverage
