### 2026-05-05 dac0629
add test for reserved keywords with limited scope
