### 2026-05-06 75a0c8b
review comments, implement escape character for enum items
