+++ tests/data/yaml2nxdl/enumerations.yaml
@@ -16,7 +16,7 @@ NXenums(NXobject):
         The items are stored in a dict with the items keyword.
       \enumeration:
         # comment 1
-        items: [value0, value1, value2]
+        \items: [value0, value1, value2]
     closed_enum_with_numbers(NX_NUMBER):
       \doc: |
         This is a closed enumeration with NX_NUMBER items.
@@ -59,7 +59,7 @@ NXenums(NXobject):
         This is an open enumeration without any docstrings for the individual items.
       \enumeration:
         \open: true
-        items: [value0, value1, value2]
+        \items: [value0, value1, value2]
     open_enum_no_docs_with_comments:
       \doc: |
         This is an open enumeration without any docstrings for the individual items.
@@ -67,7 +67,7 @@ NXenums(NXobject):
       \enumeration:
         # comment4
         \open: true
-        items: [value0, value1, value2]
+        \items: [value0, value1, value2]
     open_enum_with_docs:
       \doc: |
         This is a open enumeration, with each item having a docstring.

[…]
