### 2026-05-06 50fb672
guard definition-level keywords against missing \ prefix
