### 2026-05-06 f017c66
Apply suggestions from code review
Co-authored-by: Markus Kühbach <mkuehbach@users.noreply.github.com>
