### 2026-05-06 3afdc43
Apply suggestions from code review
Co-authored-by: Markus Kühbach <mkuehbach@users.noreply.github.com>
