+++ tests/data/roundtrip/nxdl2yml.Ref_NXcanSAS.yaml
@@ -1,2425 +1,0 @@
-category: application
-doc: |
-  Implementation of the canSAS standard to store reduced small-angle scattering data of any dimension.
-  
-  .. index:: canSAS
-  
-  For more details, see:
-  
-  * http://www.cansas.org/
-  * http://www.cansas.org/formats/canSAS1d/1.1/doc/
-  * http://cansas-org.github.io/canSAS2012/
-  * https://github.com/canSAS-org/NXcanSAS_examples
-  
-  The minimum requirements for *reduced* small-angle scattering data
-  as described by canSAS are summarized in the following figure:
-  
-  .. _canSAS_2012_minimum:
-  
-  .. figure:: canSAS/2012-minimum.png
-     :width: 60%
-     
-     The minimum requirements for *reduced* small-angle scattering data.
-     (:download:`full image <canSAS/2012-minimum.png>`)
-     See :ref:`below <NXcanSAS_minimum>` for the minimum required
-     information for a NeXus data file
-     written to the NXcanSAS specification.
-  
-  .. rubric::  Implementation of canSAS standard in NeXus
-  
-  This application definition is an implementation of the canSAS
-  standard for storing both one-dimensional and multi-dimensional
-  *reduced* small-angle scattering data.
-  
-  * NXcanSAS is for reduced SAS data and metadata to be stored together in one file.
-  * *Reduced* SAS data consists of :math:`I(\vec{Q})` or :math:`I(|\vec{Q}|)`
-  * External file links are not to be used for the reduced data.
-  * A good practice/practise is, at least, to include a reference to how the data was acquired and processed.  Yet this is not a requirement.
-  * There is no need for NXcanSAS to refer to any raw data.
-  
-  The canSAS data format has a structure similar to NeXus, not identical.
-  To allow canSAS data to be expressed in NeXus, yet identifiable
-  by the canSAS standard, an additional group attribute ``canSAS_class``
-  was introduced.  Here is the mapping of some common groups.
-  
-  ===============  ============  ==========================
-  group (*)        NX_class      canSAS_class
-  ===============  ============  ==========================
-  sasentry         NXentry       SASentry
-  sasdata          NXdata        SASdata
-  sasdetector      NXdetector    SASdetector
-  sasinstrument    NXinstrument  SASinstrument
-  sasnote          NXnote        SASnote
-  sasprocess       NXprocess     SASprocess
-  sasprocessnote   NXcollection  SASprocessnote
-  sastransmission  NXdata        SAStransmission_spectrum
-  sassample        NXsample      SASsample
-  sassource        NXsource      SASsource
-  ===============  ============  ==========================
-  
-  (*) The name of each group is a suggestion,
-  not a fixed requirement and is chosen as fits each data file.
-  See the section on defining
-  :ref:`NXDL group and field names <RegExpName>`.
-  
-  Refer to the NeXus Coordinate System drawing (:ref:`Design-CoordinateSystem`)
-  for choice and direction of :math:`x`, :math:`y`, and :math:`z` axes.
-  
-  .. _NXcanSAS_minimum:
-  
-  .. rubric:: The minimum required information for a NeXus data file
-     written to the NXcanSAS specification.
-  
-  .. literalinclude:: canSAS/minimum-required.txt
-     :tab-width: 4
-     :linenos:
-     :language: text
-
-# NOTE:
-# This NXDL refers to several image files (.jpg, .png) in its documentation.
-# All such related resources are stored in the related subdirectory: ./canSAS/
-# In general, for an NXDL file:  NXsomenxdl.nxdl.xml
-# all related resources should be stored in subdirectory:  ./somenxdl/
-type: group
-NXcanSAS(NXobject):
-  (NXentry):
-    \@default:
[…]
