+++ tests/data/nxdl2yaml/ref_doc_nxdl2yaml.yaml
@@ -1,5 +1,5 @@
-category: application
-doc: |
+\category: application
+\doc: |
   This is a definition for data to be archived by ICAT (http://www.icatproject.org/).
   
   .. text from the icatproject.org site
@@ -8,13 +8,13 @@ doc: |
           interface to all ISIS experimental data and will provide
           a mechanism to link all aspects of ISIS research from
           proposal through to publication.
-type: group
+\type: group
 NXarchive(NXobject):
   (NXentry)entry:
     \@index:
     title:
     experiment_identifier(NX_CHAR):
-      doc:
+      \doc:
       - |
         unique identifier for the experimenta
           lkokö
@@ -27,7 +27,7 @@ NXarchive(NXobject):
       - |
         This is test doc.
     experiment_description(NX_CHAR):
-      doc:
+      \doc:
       - |
         Part1: Text line 1.
         part1: Text line 2
@@ -37,31 +37,31 @@ NXarchive(NXobject):
           \term: term
           \url: url
     only_iso_string(NX_CHAR):
-      doc: |
+      \doc: |
         \xref:
           \spec: ISO 18115-1:2023
           \term: 1.1
-          \url: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:\term:1.1
+          \url: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:term:1.1
     double_iso_string(NX_CHAR):
-      doc:
+      \doc:
       - |
         Way of scanning the energy axis (fixed or sweep).
       - |
         \xref:
           \spec: ISO 18115-1:2023
           \term: 12.65
-          \url: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:\term:12.65
+          \url: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:term:12.65
       - |
         \xref:
           \spec: ISO 18115-1:2023
           \term: 12.66
-          \url: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:\term:12.66
+          \url: https://www.iso.org/obp/ui/en/#iso:std:iso:18115:-1:ed-3:v1:en:term:12.66
     collection_identifier(NX_CHAR):
-      doc: |
+      \doc: |
         ID of user or DAQ define group of data files
         Brief summary of the collection, including grouping criteria
     entry_identifier(NX_CHAR):
-      doc: |
+      \doc: |
         unique identifier for this measurement as provided by the facility
     start_time(NX_DATE_TIME):
 
@@ -93,9 +93,9 @@ NXarchive(NXobject):
 # <definition xmlns="http://definition.nexusformat.org/nxdl/3.1" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" category="application" type="group" name="NXarchive" extends="NXobject" xsi:schemaLocation="http://definition.nexusformat.org/nxdl/3.1 ../nxdl.xsd">
 #   <doc>
 #        This is a definition for data to be archived by ICAT (http://www.icatproject.org/).
-# 
+#
 #        .. text from the icatproject.org site
-# 
+#
 #        		the database (with supporting software) that provides an
 #        		interface to all ISIS experimental data and will provide
 #        		a mechanism to link all aspects of ISIS research from
@@ -109,11 +109,11 @@ NXarchive(NXobject):
 #            unique identifier for the experimenta
 #              lkokö
 #              sddfdfd
-# 
+#
[…]
