+++ tests/data/yaml2nxdl/NXtest_links.yaml
@@ -1,5 +1,6 @@
 \category: base
 \doc: Test case for verifying that the parser can handle links correctly.
+\type: group
 NXentry:
    (NXdata):
      polar_angle(link):
