### 2026-05-06 f13cd1a
Apply suggestions from code review
Co-authored-by: Markus Kühbach <mkuehbach@users.noreply.github.com>
