+++ tests/data/yaml2nxdl/NXellipsometry-docCheck.yaml
@@ -42,7 +42,7 @@
       - R. M. A. Azzam and N. M. Bashara, Ellipsometry and Polarized Light, North-Holland Publishing Company, 1977.
       - H. G. Tompkins and E. A. Irene, Handbook of Ellipsometry, William Andrew, 2005.
 
-      Open acces sources':'
+      Open access sources':'
       - https':'//www.angstromadvanced.com/resource.asp
       - https':'//pypolar.readthedocs.io/en/latest/
     definition(NX_CHAR):
