### 2026-05-06 b1c9183
do not allow blank recommended, optional, etc. keywords
