### 2026-05-06 6558b1e
clarify behavior of \open: false (discouraged)
