+++ tests/test_nyaml2nxdl.py
@@ -31,7 +31,12 @@ from click.testing import CliRunner
 from helpers import check_and_replace_latest_copyright, compare_matches
 
 from nyaml import cli as nyaml2nxdl
-from nyaml.helper import remove_namespace_from_tag
+from nyaml.helper import (
+    _YAML_ONLY_KEYWORDS,
+    LIMITED_RESERVED_KEYWORDS,
+    RESERVED_KEYWORDS,
+    remove_namespace_from_tag,
+)
 from nyaml.nyaml2nxdl import handle_each_part_doc
 
 NYAML2NXDL_DATA_DIR = Path(__file__).parent / "data" / "yaml2nxdl"
@@ -64,10 +69,9 @@ def test_nametypes_nyaml2nxdl():
     test_yml_file = str(NYAML2NXDL_DATA_DIR / "allowed_nameType.yaml")
     test_xml_file = str(NYAML2NXDL_DATA_DIR / "allowed_nameType.nxdl.xml")
     desired_matches = ["partial", "specified", "any"]
-
     compare_matches(ref_xml_file, test_yml_file, test_xml_file, desired_matches)
     os.remove(test_xml_file)
-    sys.stdout.write("Test on open/closed enumerations okay.\n")
+    sys.stdout.write("Test on nameType okay.\n")
 
 
 @pytest.mark.parametrize(
@@ -222,16 +226,14 @@ def test_nyaml2nxdl_prohibited_nameType(
 
     key, name_type_value = input_tuple
 
-    # Create the structure with the tuple data
     yaml_data = OrderedDict(
         {
-            "category": "base",
-            "doc": "Test",
-            "NXtest": OrderedDict({key: {"nameType": name_type_value}}),
+            r"\category": "base",
+            r"\doc": "Test",
+            "NXtest": OrderedDict({key: {r"\nameType": name_type_value}}),
         }
     )
 
-    # Write the YAML to a file
     test_file = Path(tmp_path) / "test.yaml"
     output_file = Path(tmp_path) / "out.nxdl.xml"
     with open(test_file, "w") as file:
@@ -243,7 +245,6 @@ def test_nyaml2nxdl_prohibited_nameType(
         [str(test_file), "--output-file", str(output_file)],
     )
 
-    # Assert that we encounter the expected error message
     assert result.exit_code == exit_code
     assert expected_error in (str(result.output) + str(result.exception))
     os.remove(test_file)
@@ -254,33 +255,15 @@ def test_file_line_error():
     """
     In this test the yaml file line in the error message is tested.
     """
-    test_yml_file = str(NYAML2NXDL_DATA_DIR / "NXfilelineError1.yaml")
-    out_nxdl = str(NYAML2NXDL_DATA_DIR / "NXfilelineError1.nxdl.xml")
-    out_yaml = str(NYAML2NXDL_DATA_DIR / "temp_NXfilelineError1.yaml")
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [test_yml_file])
-    assert result.exit_code == 1
-    assert "13" in str(result.exception)
-    os.remove(out_nxdl)
-    os.remove(out_yaml)
-
-    test_yml_file = str(NYAML2NXDL_DATA_DIR / "NXfilelineError2.yaml")
-    out_nxdl = str(NYAML2NXDL_DATA_DIR / "NXfilelineError2.nxdl.xml")
-    out_yaml = str(NYAML2NXDL_DATA_DIR / "temp_NXfilelineError2.yaml")
-    result = CliRunner().invoke(nyaml2nxdl.launch_tool, [test_yml_file])
-    assert result.exit_code == 1
[…]
