### 2026-05-07 f4dc8e6
prepare for v0.3
