### 2026-05-07 466f751
Resolve dimension ambiguity.
