### 2026-05-07 ee26094
allow \deprecated as a reserved keyword
