+++ tests/data/nxdl2yaml/Ref_NXattributes.yaml
@@ -20,6 +20,7 @@ NXellipsometry_base_draft(my_test_extends):
         documentation no. 3
     experiment_description:
       \exists: optional
+      \deprecated: true
     start_time(NX_DATE_TIME):
       \exists: optional
       \unit: NX_TIME
@@ -39,6 +40,7 @@ NXellipsometry_base_draft(my_test_extends):
       \doc: |
         documentation no. 7
       \@version:
+        \deprecated: true
     calibration_data(NX_NUMBER):
       \unit: NX_UNITLESS
       \doc: |
@@ -47,7 +49,7 @@ NXellipsometry_base_draft(my_test_extends):
         (transmission in air).
 
 # ++++++++++++++++++++++++++++++++++ SHA HASH ++++++++++++++++++++++++++++++++++
-# abafd962b1e78462ba064863826bfe6a7cbc98c73c91ffefd9c923faee03cf9e
+# 9afa2ceb0d71c7e1cb16d3098abcb36bb25e118cad2373e9d3195a54b9a534cf
 # <?xml version='1.0' encoding='UTF-8'?>
 # <?xml-stylesheet type="text/xsl" href="nxdlformat.xsl"?>
 # <!--
@@ -99,7 +101,7 @@ NXellipsometry_base_draft(my_test_extends):
 #                 documentation no. 3
 #             </doc>
 #         </field>
-#         <field name="experiment_description" optional="false"/>
+#         <field name="experiment_description" optional="false" deprecated="true"/>
 #         <field name="start_time" type="NX_DATE_TIME" optional="false" units="NX_TIME"/>
 #         <field name="program_name">
 #             <doc>
@@ -120,7 +122,7 @@ NXellipsometry_base_draft(my_test_extends):
 #             <doc>
 #                 documentation no. 7
 #             </doc>
-#             <attribute name="version"/>
+#             <attribute name="version" deprecated="true"/>
 #         </field>
 #         <field name="calibration_data" type="NX_NUMBER" units="NX_UNITLESS">
 #             <doc>
