### 2026-05-07 830c6d2
Raise error in rank and dim inconsistent in backward direction.
