### 2026-05-07 06e38e6
allow `\deprecated` as a reserved keyword (#97)
Co-authored-by: lukaspie <lukaspie@github.com>
