+++ tests/data/yaml2nxdl/NXattributes.yaml
@@ -13,6 +13,7 @@ NXellipsometry_base_draft(my_test_extends):
       \doc: documentation no. 3
     experiment_description:
       \exists: required
+      \deprecated: true
     start_time(NX_DATE_TIME):
       \exists: required
       \unit: NX_TIME
@@ -28,6 +29,7 @@ NXellipsometry_base_draft(my_test_extends):
       \exists: ['max', 5]
       \doc: documentation no. 7
       \@version:
+        \deprecated: true
     calibration_data(NX_NUMBER):
       \unit: NX_UNITLESS
       \doc: |
