### 2026-05-07 b0d4a29
Resolve dimension ambiguity. (#96)
* Resolve dimension ambiguity.

* Raise error in rank and dim inconsistent in backward direction.

* fix typo.

* fix test.
