+++ .cspell/custom-dictionary.txt
@@ -57,3 +57,4 @@ somedocstring
 someint
 somevalue
 val
+sourceline
