### 2026-05-08 483b49f
bump setuptools and setuptools-scm, add citation
