+++ docs/assets/nyaml-citation.bib
@@ -1,0 +1,8 @@
+@Article{nyaml,
+  author  = {Rubel Mozumder and Lukas Pielsticker and Markus K\"uhbach and Andrea Albino and Florian Dobener and Sherjeel Shabih and Jos\'{e} A. M\'{a}rquez Prieto and Sandor Brockhauser and Claudia Draxl and {Christoph T.} Koch and {Heiko B.} Weber},
+  journal = {Journal of Open Source Software},
+  title   = {{nyaml: Format Converter for the NeXus Data Model}},
+  year    = {2026},
+}
+
+@Comment{jabref-meta: databaseType:bibtex;}
