+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.15.11
+    rev: v0.15.12
     hooks:
       # Run the linter.
       - id: ruff-check
@@ -39,5 +39,6 @@ repos:
           - README.md
           - CITATION.cff
           - docs/**/*.md
+          - docs/**/*.bib
           - src/nyaml/**/*.py
           - tests/**/*.py
