+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=78.1.1", "setuptools-scm[toml]>=6.2"]
+requires = ["setuptools>=80", "setuptools-scm[toml]>=8"]
 build-backend = "setuptools.build_meta"
 
 [project]
