+++ .github/workflows/draft-pdf.yml
@@ -13,11 +13,12 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-
-      - name: Check spelling
-        uses: codespell-project/actions-codespell@v2
+      - name: Check Spell
+        uses: streetsidesoftware/cspell-action@v7
         with:
-          path: paper
+          files: paper/*.md
+          config: ./.cspell/cspell.config.yaml
+          strict: true
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
