### 2026-05-11 dc71faa
prepare v1.0 (#99)
Co-authored-by: lukaspie <lukaspie@github.com>
