### 2026-05-11 0b3cfa8
Fix Atom prob microscopy
