+++ paper/paper.md
@@ -95,7 +95,7 @@ In the reverse direction — **XML to YAML** conversion (see \autoref{fig:nyaml_
 The NeXus International Advisory Committee (NIAC) is the governing body responsible for overseeing the development and maintenance of the NeXus data standard. A core responsibility of the NIAC is the stewardship of the NeXus Definition Language (NXDL), the XML-based schemata that define the hierarchical structure and semantics of NeXus data files [@Könnecke:2015]. As part of its mission to standardize NeXus definitions in NXDL, the NIAC recently reviewed and formally accepted the `nyaml` tool. Following a successful evaluation, the NIAC endorsed `nyaml` as the recommended tool for preparing NeXus definition proposals. In support of this decision, the official NeXus definition repository [@nexusDef:2024] was updated to integrate `nyaml` into its workflow through the addition of two makefile targets: 'make nyaml' [@nyamlIntegration:2024], which converts existing definitions from the canonical nxdl.xml format into .yaml, and 'make nxdl', which converts modified or newly added .yaml files back into valid nxdl.xml for submission and version control. This integration ensures that contributions made in `.yaml` are compatible with the existing XML-based infrastructure. The adoption of `nyaml` by NIAC reflects an ongoing commitment to fostering community engagement and modernizing the technical tools underpinning the NeXus standard [@NIAC:2025].
 
 # Research impact statement
-The development of `nyaml` software was initiated in 2023 [@nyamlFirstCommit2023] by FAIRmat as part of its efforts to enhance the data modeling capabilities for materials science within the NeXus framework. The project has grown organically in response to NeXus evolution needs and has since received contributions from a few NeXus community members and developers from FAIRmat. Subsequently, the tool has been used in developing several data models for different techniques in experimental materials science, including Electron Microscopy (EM) [@NIACEM2025], Multi-dimensional Photoemission Spectroscopy (MPES) [@NIACMPES2025], Optical Spectroscopy [@NIACOS2025], and Atomic Probe Microscopy [@NIACAPM2025]. The `nyaml` project is open source and hosted on GitHub [@nyamlGH2023], where users and developers can contribute to its ongoing development, report issues, and suggest enhancements. Comprehensive documentation, including installation instructions, usage guidelines, and examples, is available online at [GitHub pages](https://fairmat-nfdi.github.io/nyaml), facilitating adoption by the broader NeXus community.
+The development of `nyaml` software was initiated in 2023 [@nyamlFirstCommit2023] by FAIRmat as part of its efforts to enhance the data modeling capabilities for materials science within the NeXus framework. The project has grown organically in response to NeXus evolution needs and has since received contributions from a few NeXus community members and developers from FAIRmat. Subsequently, the tool has been used in developing several data models for different techniques in experimental materials science, including Electron Microscopy (EM) [@NIACEM2025], Multi-dimensional Photoemission Spectroscopy (MPES) [@NIACMPES2025], Optical Spectroscopy [@NIACOS2025], and Atom Probe Microscopy [@NIACAPM2025]. The `nyaml` project is open source and hosted on GitHub [@nyamlGH2023], where users and developers can contribute to its ongoing development, report issues, and suggest enhancements. Comprehensive documentation, including installation instructions, usage guidelines, and examples, is available online at [GitHub pages](https://fairmat-nfdi.github.io/nyaml), facilitating adoption by the broader NeXus community.
 
 # AI usage disclosure:
 No generative AI tools were used in the design and development of this software.
