### 2026-05-12 d1950d8
Include yaml linter for CITATION.cff file.
