+++ CITATION.cff
@@ -47,15 +47,15 @@ abstract:
   nyaml is a Python software tool designed for developers and users of the NeXus data exchange format
   (https://www.nexusformat.org/). The tool converts files that follow the NeXus Definition Language
   (NXDL) from XML to [YAML](https://yaml.org/), and vice versa.
-  
+
   NXDL is used to define experiment instruments, setups, or scientific data objects from experiments
   in a schema-based structure. Writing NeXus data schemas via the indentation-based formatting of
   YAML offers a more concise writing and reduces common sources of formatting inconsistencies,
   while still adhering to NeXus standards.
-  
+
   The tool requires users to follow a specific grammar for embedding NeXus data structures to ensure
   compatibility with NXDL (see the full documentation at https://fairmat-nfdi.github.io/nyaml/).
-  
+
   The work is funded by the Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) - 460197019 (FAIRmat)
   (https://gepris.dfg.de/gepris/projekt/460197019?language=en).
 
