### 2026-05-12 3fbbd2c
Include yaml linter for CITATION.cff file. (#100)
* Include yaml linter for CITATION.cff file.

* Update the release version.
