+++ CITATION.cff
@@ -40,9 +40,9 @@ authors:
   - given-names: Heiko B.
     family-names: Weber
     orcid: 'https://orcid.org/0000-0002-6403-9022'
-doi: 10.5281/zenodo.20135563
+doi: 10.5281/zenodo.13860810
 repository-code: 'https://github.com/FAIRmat-NFDI/nyaml'
-url: https://doi.org/10.5281/zenodo.20135563
+url: https://doi.org/10.5281/zenodo.13860810
 abstract:
   nyaml is a Python software tool designed for developers and users of the NeXus data exchange format
   (https://www.nexusformat.org/). The tool converts files that follow the NeXus Definition Language
