### 2026-05-12 e825428
Make the title in CITATION.cff and README.md file same. (#101)
* Make the title in CITATION.cff and README.md file same.

* Fix DOI in REAMDME.md file.

* Back to the old doi.
