### 2026-05-13 e708791
Remove duplication of cspell file and use the common one json config file
