+++ cspell.config.yaml
@@ -1,10 +1,0 @@
-
-version: '0.2'
-dictionaryDefinitions:
-    - name: custom-dictionary
-      path: '.cspell/custom-dictionary.txt'
-      addWords: true
-dictionaries:
-    - custom-dictionary
-ignorePaths:
-    - 'paper/paper.bib'
