+++ .github/workflows/draft-pdf.yml
@@ -17,7 +17,7 @@ jobs:
         uses: streetsidesoftware/cspell-action@v7
         with:
           files: paper/*.md
-          config: ./.cspell/cspell.config.yaml
+          config: ./cspell.json
           strict: true
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
