### 2026-05-13 8bfbb1b
Removed paper draft github action.
