+++ .github/workflows/draft-pdf.yml
@@ -1,35 +1,0 @@
-name: Draft PDF
-
-on:
-  push:
-    paths:
-      - paper/**
-      - .github/workflows/draft-pdf.yml
-
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: Paper Draft
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Check Spell
-        uses: streetsidesoftware/cspell-action@v7
-        with:
-          files: paper/*.md
-          config: ./cspell.json
-          strict: true
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          # This should be the path to the paper within your repo.
-          paper-path: paper/paper.md
-      - name: Upload
-        uses: actions/upload-artifact@v4
-        with:
-          name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
-          path: paper/paper.pdf
