### 2026-05-13 dd64cc4
JOSS paper (#71)
* First commit.

* JOSS paper: summary and statement of needs (#77)

Co-authored-by: RubelMozumder <32923026+RubelMozumder@users.noreply.github.com>

* add a CI/CD workflow for building the paper draft (#72)

* joss NIAC (#75)

* rebase

* Update paper.md

* Update paper.md

---------

Co-authored-by: Rubel <rubel.mozumder@outlook.com>

* Bring all together.

* draft

* reorder the authors.

update paper and figure.

Update with build.sh file.

Update.

minor update.

* Fix missing reference, check consistency for keyword and fix brocken author in reference.

* Little update o
[…]
