+++ paper/paper.md
@@ -1,0 +1,107 @@
+---
+title: "nyaml: Format Converter for the NeXus Data Model"
+tags:
+  - Python
+  - NeXus
+  - NXDL
+  - YAML
+  - XML
+
+authors:
+  - given-names: Rubel
+    surname: Mozumder
+    orcid: 0009-0007-5926-6646
+    affiliation: 1
+  - given-names: Lukas
+    surname: Pielsticker
+    orcid: 0000-0001-9361-8333
+    affiliation: "1,2"
+  - given-names: Markus
+    surname: Kühbach
+    orcid: 0000-0002-7117-5196
+    affiliation: 1
+  - given-names: Andrea
+    surname: Albino
+    orcid: 0000-0001-9280-7431
+    affiliation: 1
+  - given-names: Florian
+    surname: Dobener
+    orcid: 0000-0003-1987-6224
+    affiliation: 1
+  - given-names: Sherjeel
+    surname: Shabih
+    orcid: 0009-0008-6635-4465
+    affiliation: 1
+  - given-names: José A.
+    surname: Márquez Prieto
+    orcid: 0000-0002-8173-2566
+    affiliation: 1
+  - given-names: Sandor
+    surname: Brockhauser
+    orcid: 0000-0002-9700-4803
+    affiliation: 1
+  - given-names: Claudia
+    surname: Draxl
+    orcid: 0000-0003-3523-6657
+    affiliation: 1
+  - given-names: Christoph
+    surname: Koch
+    orcid: 0000-0002-3984-1523
+    affiliation: 1
+  - given-names: Heiko B.
+    surname: Weber
+    orcid: 0000-0002-6403-9022
+    affiliation: 3
+
+affiliations:
+  - name: Physics Department and CSMB, Humboldt-Universität zu Berlin, Zum Großen Windkanal 2, D-12489 Berlin, Germany
+    index: 1
+  - name: Department Heterogeneous Reactions, Max Planck Institute for Chemical Energy Conversion, Stiftstraße 34-36, D-45470 Mülheim an der Ruhr, Germany
+    index: 2
+  - name: Lehrstuhl für Angewandte Physik, Friedrich-Alexander-Universität Erlangen-Nürnberg, Staudtstraße 7, D-91058 Erlangen, Germany
+    index: 3
+
+date: 01 July 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+The NeXus scientific data format standard [@Könnecke:2015; @Könnecke2006; @Klosowski1997], which was originally introduced for neutron, X-ray, and muon science, has in recent years seen a significant enhancement across diverse scientific domains such as materials science. NeXus definitions—comprising application definitions and base classes—describe the hierarchical structure and semantics of valid NeXus files. They are written in XML [@xml10] using the NeXus Definition Language (NXDL), which is itself defined in XSD (XML Schema Definition) [@w3c:xmlschema1].
+
+`nyaml` is a Python-based tool with both a command-line interface and an application programming interface (API) that facilitates the conversion between NXDL XML and a simplified YAML [@YAML11] representation. YAML's indentation-based syntax enhances human readability and simplifies manual editing. By providing a reliable, lossless round-trip conversion between XML and YAML, `nyaml` enables developers to edit NeXus definitions efficiently without sacrificing structural or semantic fidelity.
+
+# Statement of need
+
[…]
