+++ paper/paper.bib
@@ -1,0 +1,174 @@
+@software{PyYAML:2024,
+  author = {{YAML and Python communities}},
+  title = {PyYAML},
+  year = {2024},
+  month = {Aug},
+  version = {latest},
+  url = {https://github.com/yaml/pyyaml},
+}
+@software{Click:2025,
+    author       = {{The click development team}},
+    title        = {Command Line Interface Creation Kit},
+    month        = {May},
+    year         = {2025},
+    version      = {latest},
+    url          = {https://github.com/pallets/click}
+}
+@software{lxml:2025,
+    author = {{Behnel, S. et al.}},
+    title = {lxml},
+    year = {2025},
+    month = {Jan},
+    version = {latest},
+    url = {https://github.com/lxml/lxml}
+}
+@article{Könnecke:2015,
+author = "K{\"{o}}nnecke, Mark and Akeroyd, Frederick A. and Bernstein, Herbert J. and Brewster, Aaron S. and Campbell, Stuart I. and Clausen, Bj{\"{o}}rn and Cottrell, Stephen and Hoffmann, Jens Uwe and Jemian, Pete R. and M{\"{a}}nnicke, David and Osborn, Raymond and Peterson, Peter F. and Richter, Tobias and Suzuki, Jiro and Watts, Benjamin and Wintersberger, Eugen and Wuttke, Joachim",
+title = "{The NeXus data format}",
+journal = "Journal of Applied Crystallography",
+year = "2015",
+volume = "48",
+number = "1",
+pages = "301--305",
+month = "Feb",
+doi = {10.1107/S1600576714027575},
+url = {https://doi.org/10.1107/S1600576714027575}
+}
+@misc{NIAC:2025,
+  title = {NIAC and NeXus Documentation},
+  author = {{NeXus International Advisory Committee}},
+  year = {2025},
+  month = {Aug},
+  url = {https://www.nexusformat.org},
+  note = {Accessed: 2025-08-01}
+}
+@techreport{YAML11,
+  author = {Ben-Kiki, Oren and Evans, Clark and Ingerson, Brian},
+  institution = {YAML.org},
+  year = 2005,
+  month = {1},
+  title = {{YAML} Ain't Markup Language ({YAML}) (tm) Version 1.1},
+  url = {https://yaml.org/spec/1.1/}
+}
+@misc{xml10,
+  author       = {Tim Bray and Jean Paoli and C. M. Sperberg-McQueen and Eve Maler and François Yergeau},
+  title        = {Extensible Markup Language (XML) 1.0 (Fifth Edition)},
+  howpublished = {\url{https://www.w3.org/TR/2008/REC-xml-20081126/}},
+  year         = {2008},
+  note         = {W3C Recommendation, World Wide Web Consortium (W3C)},
+}
+@misc{pip2025,
+  author       = {{The Python Packaging Authority}},
+  title        = {pip: The Python Package Installer},
+  year         = {2025},
+  howpublished = {\url{https://pip.pypa.io/}},
+  note         = {Accessed: 2025-07-15}
+}
+@article{Klosowski1997,
+title = {NeXus: A common format for the exchange of neutron and synchroton data},
+journal = {Physica B: Condensed Matter},
+volume = {241-243},
+pages = {151-153},
+year = {1997},
+note = {Proceedings of the International Conference on Neutron Scattering},
+issn = {0921-4526},
+doi = {10.1016/S0921-4526(97)00865-X},
+url = {https://www.sciencedirect.com/science/article/pii/S092145269700865X},
[…]
