### 2026-06-15 9b92c5b
update bib file with paper release metadata
