+++ docs/assets/nyaml-citation.bib
@@ -1,8 +1,15 @@
 @Article{nyaml,
-  author  = {Rubel Mozumder and Lukas Pielsticker and Markus K\"uhbach and Andrea Albino and Florian Dobener and Sherjeel Shabih and Jos\'{e} A. M\'{a}rquez Prieto and Sandor Brockhauser and Claudia Draxl and {Christoph T.} Koch and {Heiko B.} Weber},
-  journal = {Journal of Open Source Software},
-  title   = {{nyaml: Format Converter for the NeXus Data Model}},
-  year    = {2026},
+  author    = {Rubel Mozumder and Lukas Pielsticker and Markus Kühbach and Andrea Albino and Florian Dobener and Sherjeel Shabih and José A. Márquez Prieto and Sandor Brockhauser and Claudia Draxl and Christoph Koch and {Heiko B.} Weber},
+  journal   = {Journal of Open Source Software},
+  title     = {{nyaml: Format Converter for the NeXus Data Model}},
+  year      = {2026},
+  issn      = {2475-9066},
+  month     = {June},
+  number    = {122},
+  pages     = {10231},
+  volume    = {11},
+  doi       = {10.21105/joss.10231},
+  publisher = {The Open Journal},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
